### 2019-12-12 #16 @megies
ObsPy: Software Submission for Review
Submitting Author: @megies
Package Name: ObsPy
One-Line Description of Package: A Python Toolbox for seismology/seismological observatories.
Repository Link:  https://github.com/obspy/obspy
Version submitted:   [1.1.1](https://github.com/obspy/obspy/releases/tag/1.1.1)
Version accepted: TBD

---

ObsPy is an open-source project dedicated to provide a **Python framework for processing seismological** data. It provides parsers for common file formats, clients to access data centers and seismological signal processing routines which allow the manipulation of seismological time series (see [Beyreuther et al. 2010](http://www.seismosoc.org/publications/SRL/SRL_81/srl_81-3_es/), [Megies et al. 2011](http://www.annalsofgeophysics.eu/index.php/annals/article/view/4838), [Krischer et al. 2015](http://iopscience.iop.org/article/10.1088/1749-4699/8/1/014003)).

The goal of the ObsPy project is to facilitate **rapid application development for seismology**.

ObsPy is licensed under the GNU Lesser General Public License (LGPL) v3.0.

- Please indicate which [category or categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html) this package falls under:
	- [ ] Data visualization*  <-- didn't tick this, because the benefit of a 'pre-submission inquiry' didn't become obvious for me, the issue form seems to be largely the same

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [this section](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html#notes-on-categories) of our guidebook.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):

It falls under these categories, because we retrieve and munge data. Also loads of the functionality is concerning geospatial aspects (can't be avoided in seismology). Being as high-level as we are in functionality, I'd also consider it to be educational, since you can illustrate how to do many processing steps (and what pitfalls there are)  in a few lines of code.
If this is more about the core aspects of the software, I guess you could also argue for only `retrieval` and `munging`, and the other are rather side aspects.. although it's unclear to me in which bullet item signal processing is included (`munging` I guess?).

Mainly scientists/researchers, but we know that obspy is also used in industry (no idea about numbers, though). Scientific applications span most fields of seismology that involve real or synthetic data, basically only excluding purely theoretical, pen and paper seismology.

There are various packages with I/O of up to a handful file formats at a time, usually concerned only with 1-3 file formats used by the authors. I am not aware of other packages that offer access clients for all important protocols like we do (FDSN web services, SeedLink, Earthworm, Arclink, NRL, syngine, ...). Most important point would be the widespread use in projects built on top of obspy, most likely.

no pre-submission enquiry filed.

For details about the pyOpenSci packaging requirements, see our [packaging guide](https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html). Confirm each of the following by checking the box.  This package:

    - this feels a bit unclear, but in any case I'm not aware of any ToSs we would violate
- [x] has an [OSI approved license](https://opensource.org/licenses)
    - LGPLv3
    - these instructions are in the wiki which is linked in the README to keep it more concise
    - docs.obspy.org built from source with sphinx
    - submodule API landing pages (i.e. `__init__.py`s) hold the submodules' most important functions with explanations and usage examples,
    - see http://tests.obspy.org/
    - Travis for Linux + Mac tests
    - Appveyor for Windows tests
    - Circle for quick linting and Flake8

We have citable papers out, so no need for this I think. If anything we might do a "history of obspy" short story?

- [ ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.
- [ ] The package is not a "minor utility" as defined by JOSS's [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements): "Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.

- [x] I agree to abide by [pyOpenSci's Code of Conduct](https://www.pyopensci.org/dev_guide/peer_review/coc.html) during the review process and in maintaining my package should it be accepted.

We're using [Contributor Covenant Code of Conduct v1.4](https://github.com/obspy/obspy/blob/master/CODE_OF_CONDUCT.md)

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*

[Editor and review templates can be found here](https://www.pyopensci.org/dev_guide/appendices/templates.html)
