### 2019-12-12 #16 @megies
Code of conduct was added after last stable release, it's in `master` and will be part of soon-to-be-released 1.2.0.
