### 2019-12-12 #16 @megies
> are there people who you could suggest as reviewers?

That'd be somebody working in seismonology, good in Python and not using obspy? Is that even possible?

:point_up_2: just kidding :stuck_out_tongue_winking_eye: but yeah, most people I know would be biased :wink:
