### 2019-12-12 f32baf7
stationxml1.1: more info on URI in docstrings for source_id + warn on invalid URI
