+++ obspy/core/inventory/station.py
@@ -114,7 +114,9 @@ class Station(BaseNode):
             underwater sites, where 0 is sea level. (schema version >=1.1)
         :type source_id: str, optional
         :param source_id: A data source identifier in URI form
-           (schema version >=1.1)
+            (schema version >=1.1). URIs are in general composed of a 'scheme'
+            and a 'path' (optionally with additional components), the two of
+            which separated by a colon.
         """
         self.latitude = latitude
         self.longitude = longitude
