### 2019-12-12 fada1fe
stationxml1.1: add proper setter method for Identifiers
warn on invalid URIs and some docstring improvements
