### 2019-12-12 3b6d79c
stationxml1.1: make sure equipments get stored as a list if an actual iterator was passed in
