+++ obspy/core/inventory/station.py
@@ -254,10 +254,13 @@ class Station(BaseNode):
         if not hasattr(value, "__iter__"):
             msg = "equipments needs to be an iterable, e.g. a list."
             raise ValueError(msg)
-        if any([not isinstance(x, Equipment) for x in value]):
+        # make sure to unwind actual iterators, or the just might get exhausted
+        # at some point
+        equipments = [equipment for equipment in value]
+        if any([not isinstance(x, Equipment) for x in equipments]):
             msg = "equipments can only contain Equipment objects."
             raise ValueError(msg)
-        self._equipments = value
+        self._equipments = equipments
         # if value is None or isinstance(value, Equipment):
         #    self._equipment = value
         # elif isinstance(value, dict):
