### 2019-12-12 ff1f976
stationxml1.1: get rid of trivial getter/setter for "subject" on Comment
