### 2019-12-12 c274d75
stationxml1.1: add setter for deprecated equipment in Channel with deprecation warning message
