### 2019-12-12 2786d39
stationxml1.1: only allow one agency for operator
include deprecation warnings for old kwarg names and setters/getters
