### 2019-12-12 8c5a229
stationxml1.1: refactoring of operators setters, remove xml-esque parts of docstring description
