### 2019-12-12 4cf9142
stationxml1.1: get rid of "double" which doesnt do anything
