### 2019-12-12 b60edc4
stationxml1.1: ignore storage_format and show warningswith info
