### 2019-12-17 60c5542
stationxml1.1: add measurement_method to custom floats/complex numbers
