### 2019-12-17 5494f9a
stationxml1.1: loads of changes to DataAvailability(Span)
which was not really implemented like described by the StationXML
schema, e.g. extent start/end can be empty and time spans were missing
completely
