### 2019-12-17 c1da414
stationxml1.1: parse new (optional) "number" attribute on Numerator/Denominator
