### 2019-12-17 6ec7455
Geographic select of inventory/network/station
Add geographic options to [Inventory,Network,Station].select(), following
the same syntax of Client.get_stations().

Also add documentation to "sampling_rate" parameter.
