+++ CHANGELOG.txt
@@ -61,6 +61,7 @@ master:
    * Fix a bug that was causing an exception being raised in `Response.plot()`
      because a float was being passed down to numpy.linspace as length of array
      (see #2533)
+   * Geographic select of inventory/network/station (see #2515)
  - obspy.clients.fdsn:
    * Add new `_discover_services` boolean flag to the Client, which allows the
      Client to skip the initial services query at instantiation.  This can

[…]
