### 2019-12-17 52f890a
Fix a typo in get_stations() and get_events() docs
