### 2019-12-18 ce58e79
stationxml1.1: deprecation warning when setting unit on Numerator/Denominator
apparently should never been possible and was removed with stationxml
1.1

see
https://github.com/FDSN/StationXML/commit/e43e5145c0c6c7e328468ce774ac9f8cada0e7dc#r36468475
