### 2019-12-18 b8b1d85
docs: add info on import statements style
