### 2019-12-18 eb27553
stationxml1.1: fix reading/writing Identifers
