### 2019-12-18 b2256ec
stationxml1.1: do not write deprecated "unit" for Numerator/Denominator
