### 2019-12-18 6d638f9
stationxml1.1: don't write decimation for Polynomial Stages
