### 2019-12-18 41354ce
stationxml1.1: adjust the testing for agency/agencies on Operator
