### 2019-12-18 96b2025
stationxml1.1: when writing: no need to add availability schema location anymore
its been included into main schema
