### 2019-12-18 dd94873
stationxml1.1: got rid of checking for xsi namespace alias being specified by user
user should be responsible for not overwriting a common namespace alias
and we don't add xsi namespace anymore since availability is now
included in main schema, so we don't add schemaLocation attributes on
root tag anymore, see previous commits
