### 2019-12-18 ff2d0b8
stationxml1.1: store "format" tag from SC3ML/Arclink Inventory XML as foreign info in extra
since it was removed in StationXML 1.1 and we don't have a canonical
place in Channel anymore.
