### 2019-12-18 5dd6150
Cleanup Sphinx comments
