### 2019-12-19 50005f3
Cope with missing pick-info
