### 2019-12-19 a75e195
Add changelog entry
