### 2019-12-19 85e810e
doc fix in Stream._groupby
