### 2019-12-19 f17ba0b
Remove verbose option and raise a useful error on fail
