+++ obspy/io/nordic/tests/test_nordic.py
@@ -58,9 +58,13 @@ class TestNordicMethods(unittest.TestCase):
         test_cat += test_event
         # Check the read-write s-file functionality
         with TemporaryWorkingDirectory():
-            sfile = _write_nordic(test_cat[0], filename=None, userid='TEST',
-                                  evtype='L', outdir='.', wavefiles='test',
-                                  explosion=True, overwrite=True)
+            with warnings.catch_warnings():
+                # Evaluation mode mapping warning
+                warnings.simplefilter('ignore', UserWarning)
+                sfile = _write_nordic(
+                    test_cat[0], filename=None, userid='TEST', evtype='L',
+                    outdir='.', wavefiles='test', explosion=True,
+                    overwrite=True)
             self.assertEqual(readwavename(sfile), ['test'])
             read_cat = Catalog()
             # raises "UserWarning: AIN in header, currently unsupported"
@@ -165,52 +169,66 @@ class TestNordicMethods(unittest.TestCase):
         test_cat.append(full_test_event())
         with self.assertRaises(NordicParsingError):
             # Raises error due to multiple events in catalog
-            _write_nordic(test_cat, filename=None, userid='TEST',
-                          evtype='L', outdir='.',
-                          wavefiles='test', explosion=True,
-                          overwrite=True)
+            with warnings.catch_warnings():
+                # Evaluation mode mapping warning
+                warnings.simplefilter('ignore', UserWarning)
+                _write_nordic(test_cat, filename=None, userid='TEST',
+                              evtype='L', outdir='.',
+                              wavefiles='test', explosion=True,
+                              overwrite=True)
         with self.assertRaises(NordicParsingError):
             # Raises error due to too long userid
-            _write_nordic(test_ev, filename=None, userid='TESTICLE',
-                          evtype='L', outdir='.',
-                          wavefiles='test', explosion=True,
-                          overwrite=True)
+            with warnings.catch_warnings():
+                warnings.simplefilter('ignore', UserWarning)
+                _write_nordic(test_ev, filename=None, userid='TESTICLE',
+                              evtype='L', outdir='.',
+                              wavefiles='test', explosion=True,
+                              overwrite=True)
         with self.assertRaises(NordicParsingError):
             # Raises error due to unrecognised event type
-            _write_nordic(test_ev, filename=None, userid='TEST',
-                          evtype='U', outdir='.',
-                          wavefiles='test', explosion=True,
-                          overwrite=True)
+            with warnings.catch_warnings():
+                warnings.simplefilter('ignore', UserWarning)
+                _write_nordic(test_ev, filename=None, userid='TEST',
+                              evtype='U', outdir='.',
+                              wavefiles='test', explosion=True,
+                              overwrite=True)
         with self.assertRaises(NordicParsingError):
             # Raises error due to no output directory
-            _write_nordic(test_ev, filename=None, userid='TEST',
-                          evtype='L', outdir='albatross',
-                          wavefiles='test', explosion=True,
-                          overwrite=True)
+            with warnings.catch_warnings():
+                warnings.simplefilter('ignore', UserWarning)
+                _write_nordic(test_ev, filename=None, userid='TEST',
+                              evtype='L', outdir='albatross',
+                              wavefiles='test', explosion=True,
+                              overwrite=True)
         invalid_origin = test_ev.copy()
 
         invalid_origin.origins = []
         with self.assertRaises(NordicParsingError):
-            _write_nordic(invalid_origin, filename=None, userid='TEST',
-                          evtype='L', outdir='.',
-                          wavefiles='test', explosion=True,
-                          overwrite=True)
[…]
