### 2019-12-20 8829ba3
make pytest run doctests by default
