+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 test_option=conftest.py
+addopts = --doctest-modules
