### 2019-12-20 0b25410
stationxml1.1: add a test reading the old fully random SttionXML 1.0
