+++ obspy/io/stationxml/tests/test_stationxml.py
@@ -29,6 +29,72 @@ from lxml import etree
 import obspy.io.stationxml.core
 
 
+fullrandom_stationxml_1_0_changes = {
+    1: b'<FDSNStationXML xmlns="http://www.fdsn.org/xml/station/1" '
+       b'schemaVersion="1.0">',
+    502: b'                            <Numerator unit="IoDwslh8kHJBTY6U28" '
+         b'plusError="-1096718463.14" '
+         b'minusError="1126505295.72">643096860.28</Numerator>',
+    503: b'                            <Numerator '
+         b'unit="SowFov80sfXY29yH8QKDkM46" plusError="1487660393.16" '
+         b'minusError="1518618020.33">-1135677195.92</Numerator>',
+    504: b'                            <Denominator '
+         b'unit="R0-VApVbeaaxNu08Vz2Yn2U" plusError="-1418173713.21" '
+         b'minusError="-696602068.55">1504792438.09</Denominator>',
+    505: b'                            <Denominator unit="YZJt" '
+         b'plusError="587099516.91" '
+         b'minusError="618144513.11">1711947659.78</Denominator>',
+    1574: b'                            <Numerator unit="UJ60-cTOQW-" '
+          b'plusError="-1762674286.27" '
+          b'minusError="-1269852413.17">-191637952.97</Numerator>',
+    1575: b'                            <Numerator '
+          b'unit="dWVUbVRdoW84E9aMmhoPMf0_Vz5" plusError="1954449462.37" '
+          b'minusError="221794746.01">-661293296.04</Numerator>',
+    1576: b'                            <Denominator unit="rGnLumgUQBtR" '
+          b'plusError="-1789269251.46" '
+          b'minusError="-2058655199.36">1333976063.09</Denominator>',
+    1577: b'                            <Denominator '
+          b'unit="VjFL7CjQnMjn.gRbuy" plusError="1009129881.95" '
+          b'minusError="487439513.87">-94223624.85</Denominator>',
+    3041: b'                            <Numerator unit="KtMY_H" '
+          b'plusError="1980779412.19" '
+          b'minusError="1967022575.02">1495868669.19</Numerator>',
+    3042: b'                            <Numerator '
+          b'unit="Tmb9fYQNp.BIIG3WpTN6n4o" plusError="97119083.65" '
+          b'minusError="-405991834.0">1874295789.18</Numerator>',
+    3043: b'                            <Denominator unit="Ium-gE2" '
+          b'plusError="-852864606.7" '
+          b'minusError="317433598.29">-1136352379.95</Denominator>',
+    3044: b'                            <Denominator '
+          b'unit="_3xqScYZeK4IifBd3hEoC-n-TZQ" plusError="-1568478443.12" '
+          b'minusError="-487684414.86">691232466.85</Denominator>',
+    }
+# don't reinsert all the changes made in #2510, it's enough to add every type
+# of change once
+fullrandom_stationxml_1_0_insertions = [
+    (216, b'                <Agency>UIyHdtVpO</Agency>'),
+    (411, b'                <StorageFormat>WwXQ2n1i</StorageFormat>'),
+    (600, b'                        <Decimation>'),
+    (600, b'                            <InputSampleRate unit="HERTZ" '
+          b'plusError="59663358.91" '
+          b'minusError="-1077596674.88">-2002635152.29</InputSampleRate>'),
+    (600, b'                            <Factor>240803815</Factor>'),
+    (600, b'                            <Offset>-262793817</Offset>'),
+    (600, b'                            <Delay unit="XPpuXfwz-L8fb_ckRH" '
+          b'plusError="1934268189.45" '
[…]
