### 2019-12-20 6450b38
stationxml1.1: suppress some deprecation warnings when reading a StationXML 1.0 file
