### 2019-12-20 223f0b4
testing: add context manager to catch and assert warnings
inheriting from and extending warnings.catch_warnings
