### 2019-12-20 68a2f0b
stationxml1.1: properly catch and assert warnings during tests
