### 2019-12-23 2e245f5
First attempt to have sphinx skip some "pages not in toctree"
