+++ misc/docs/source/packages/obspy.core.inventory.rst
@@ -1,3 +1,5 @@
+:orphan:
+
 .. currentmodule:: obspy.core.inventory
 .. automodule:: obspy.core.inventory
 
