### 2019-12-23 a6bef64
Change description of tuple parameter
