### 2019-12-25 39484eb
stationxml: work around empty string tags getting converted to a string value "None"
