### 2019-12-25 bdeb822
docs: update coding style on imports
