### 2019-12-25 ae30e69
docs: try to suppress arclink warning during docs build
