### 2020-01-06 897d27a
Added CHANGELOG details for ticket #2296 changes.
