### 2020-01-07 11b474f
changelog cleanup and rewording
