### 2020-01-09 a529106
Check if an object is inside a geographic bound
Utility function to check whether an object is within a given
latitude and/or longitude range, or within a given distance range
from a reference geographic point.
