### 2020-01-10 23927a8
deb: update Raspbian 10 buster Dockerfile
somehow keyring wasn't properly copied into the docker image no more, so
fix that on the fly
see #2530
