+++ misc/docker/README.md
@@ -210,13 +210,16 @@ $ docker push obspy/base-images:${DISTRO_FULL}
 
 To create a Raspbian (Debian ``armhf`` for Raspberry Pi) docker base image (using ``qemu``, https://wiki.debian.org/ArmHardFloatChroot).
 This imports the public key for the Raspbian repository at the start, so the package integrity can be verified later on.
+The newer Dockerfiles were updated to use ``wget`` and ``gpg`` to import
+Raspbian archive public keys since debootstrap did not copy the keyring file
+from host to image it seems, so add option ``--include=wget,gpg``.
 
 ```bash
 $ cd /tmp
 $ wget https://archive.raspbian.org/raspbian.public.key -O - | sudo gpg --import -
 $ DISTRO=stretch
 $ DISTRO_FULL=debian_9_stretch_armhf
-$ sudo qemu-debootstrap --arch=armhf --keyring /root/.gnupg/pubring.kbx ${DISTRO} ${DISTRO_FULL} http://archive.raspbian.org/raspbian 2>&1 | tee ${DISTRO_FULL}.debootstrap.log
+$ sudo qemu-debootstrap --arch=armhf --include=wget,gpg --keyring /root/.gnupg/pubring.kbx ${DISTRO} ${DISTRO_FULL} http://archive.raspbian.org/raspbian 2>&1 | tee ${DISTRO_FULL}.debootstrap.log
 $ sudo tar -C ${DISTRO_FULL} -c . | docker import - obspy/base-images:${DISTRO_FULL}
 $ docker login  # docker hub user needs write access to "obspy/base-images" of organization "obspy"
 $ docker push obspy/base-images:${DISTRO_FULL}
