### 2020-01-12 5eb20df
Select traces in Stream based on an Inventory
Select traces in Stream based on the content of an Inventory object:
trace will be selected if the inventory contains a matching channel
active at the trace start time.
