### 2020-01-15 432660e
Cleaned up accessing of instrument register values as per suggestions of @QuLogic
