### 2020-01-20 79b9674
io.seiscomp: dont set "None" for unknown empty unit name but rather empty string which seems more appropriate
