+++ CHANGELOG.txt
@@ -17,6 +17,8 @@ master:
    * fixed UTCDateTime(..., julday=366) for non-leap years. This was returning
      January 1st of the next year in case of non-leap years being used. Now it
      properly raises an out-of-bounds ValueError (see #2369)
+   * When reading StationXML/SC3ML, make sure to properly read empty string
+     fields as empty strings instead of "None" (see #2519 and #2527)
    * better ISO8601 detection for UTCDateTime objects and UTCDateTime(...,
      iso8601=False) now completely disables ISO8601 handling (see #2447)
    * Added replace method to UTCDateTime class (see #2077).
@@ -138,6 +140,11 @@ master:
    * Very large performance improvement reading large sc3ml inventory files by
      pre-indexing sensors, dataloggers and responses and reducing lxml calls
      (see #2296).
+   * When reading StationXML/SC3ML, make sure to properly read empty string
+     fields as empty strings instead of "None" (see #2519 and #2527)
+ - obspy.io.stationxml:
+   * When reading StationXML/SC3ML, make sure to properly read empty string
+     fields as empty strings instead of "None" (see #2519 and #2527)
  - obspy.io.quakeml:
    * Allow writing invalid ids but raise a warning
      (see #2104, #2090, #2093, #1872).
