### 2020-01-20 649815c
avoid passing a float to numpy.linspace "num" kwarg
raises an error on newer numpy
