### 2020-01-20 142fcee
avoid passing a float to get_evalresp_response as nfft
its fixed inside that routine as well, but anyway nfft should be int
