### 2020-01-21 58f5719
mseed: regression test for endless loop reading a fullseed file without data records
