### 2020-01-21 9dd8c5b
mseed: fix a bug trying to read a full SEED without data records as MSEED
resulting in an infinite loop
