### 2020-01-21 c7328dd
xseed: minor kwarg adjustment to StationXML 1.1
