+++ obspy/io/xseed/core.py
@@ -363,7 +363,7 @@ def _parse_to_inventory_object(p, skip_invalid_responses=True):
                 sensor=sensor,
                 pre_amplifier=None,
                 data_logger=None,
-                equipment=None,
+                equipments=None,
                 response=None,
                 description=None,
                 comments=None,
