### 2020-01-21 c47c04f
testing: adjust one test case a little further for win
