### 2020-01-22 8852ddf
stationxml: add full random v1.0 test file
since we need to add newly introduced fields in the 1.1 file and
housekeeping the diffs is too much hassle
