### 2020-01-23 25c595a
stationxml 1.1: fix order of some tags when writing stationxml
