### 2020-01-23 e7f4b79
stationxml1.1: some typo fixes in docstrings and code
