### 2020-01-23 ac3266a
inventory: make sure data availability spans is correctly set as an empty list if no spans are specified
