### 2020-01-23 1cbd0a7
inventory: correctly cast uncertainties on float types to float
see #2537
