### 2020-01-23 15dc78e
stationxml: cast to correct types when reading data availability
esp. casts of time objects to UTCDateTime
