### 2020-01-23 cae0a3a
stationxml1.1: properly write out data availability info
