### 2020-01-23 3d83ed8
stationxml1.1 tests: make sure all new additions in v1.1 are tested
also fix an old test case for data availability items which used to be
real sloppy not casting to appropriate objects..
