+++ obspy/io/stationxml/tests/test_stationxml.py
@@ -22,9 +22,11 @@ import unittest
 import warnings
 
 import obspy
+from obspy import UTCDateTime
 from obspy.core.util import AttribDict, CatchAndAssertWarnings
 from obspy.core.util.deprecation_helpers import ObsPyDeprecationWarning
 from obspy.core.inventory import (Inventory, Network, ResponseStage)
+from obspy.core.inventory.util import DataAvailability
 from obspy.core.util.base import NamedTemporaryFile
 from lxml import etree
 import obspy.io.stationxml.core
@@ -213,6 +215,88 @@ class StationXMLTestCase(unittest.TestCase):
 
         self._assert_station_xml_equality(file_buffer,
                                           expected_xml_file_buffer)
+
+        # test some new fields added in StationXML 1.1 specifically
+        net = inv[0]
+        sta = net[0]
+        cha = sta[0]
+        self.assertEqual(len(net.identifiers), 2)
+        self.assertEqual(net.identifiers[0], "abc:def")
+        self.assertEqual(net.identifiers[1], "uvw:xyz")
+        self.assertEqual(len(net.operators), 2)
+        self.assertEqual(net.operators[0].agency, 'ABC0.oszQNsC4l66ieQFM')
+        # accessing "agencies" this should raise a DeprecationWarning but its
+        # tested in another test case already
+        with CatchAndAssertWarnings():
+            self.assertEqual(net.operators[0].agencies,
+                             ['ABC0.oszQNsC4l66ieQFM'])
+        self.assertEqual(net.operators[0].contacts[0].names[0], 'A')
+        # check WaterLevel tag
+        self.assertEqual(sta.water_level, 250.4)
+        self.assertEqual(sta.water_level.lower_uncertainty, 2.3)
+        self.assertEqual(sta.water_level.upper_uncertainty, 4.2)
+        self.assertEqual(sta.water_level.unit, 'METERS')
+        self.assertEqual(cha.water_level, 631.2)
+        self.assertEqual(cha.water_level.lower_uncertainty, 5.3)
+        self.assertEqual(cha.water_level.upper_uncertainty, 3.2)
+        self.assertEqual(cha.water_level.unit, 'METERS')
+        # multiple equipments allowed now on channel, deprecation for old
+        # single equipment attribute
+        self.assertEqual(len(cha.equipments), 2)
+        self.assertEqual(cha.equipments[1].type, "some type")
+        msg = (r"Attribute 'equipment' \(holding a single Equipment\) is "
+               r"deprecated in favor of 'equipments' which now holds a list "
+               r"of Equipment objects \(following changes in StationXML 1.1\) "
+               r"and might be removed in the future. Returning the first "
+               r"entry found in 'equipments'.")
+        with CatchAndAssertWarnings(
+                clear=['obspy.core.inventory.channel'],
+                expected=[(ObsPyDeprecationWarning, msg)]):
+            self.assertEqual(cha.equipment.type, cha.equipments[0].type)
+        # check new measurementMethod attributes
+        self.assertEqual(sta.latitude.measurement_method, "GPS")
+        self.assertEqual(sta.longitude.measurement_method, "GPS")
+        self.assertEqual(sta.elevation.measurement_method,
+                         "digital elevation model")
+        self.assertEqual(cha.azimuth.measurement_method,
+                         "fibre optic gyro compass")
+        # check data availability tags
+        self.assertEqual(
+            net.data_availability.start, UTCDateTime(2011, 2, 3, 4, 5, 6))
+        self.assertEqual(
+            net.data_availability.end, UTCDateTime(2011, 3, 4, 5, 6, 7))
[…]
