### 2020-01-24 b8d64aa
fix changelog typo
