### 2020-01-24 ca022a9
check geographic filters if no filter criteria are specified
