+++ obspy/core/inventory/station.py
@@ -464,8 +464,9 @@ class Station(BaseNode):
                 minlongitude=minlongitude, maxlongitude=maxlongitude,
                 latitude=latitude, longitude=longitude, minradius=minradius,
                 maxradius=maxradius)
-            if not inside_geobounds(cha, **geo_filters):
-                continue
+            if any(value is not None for value in geo_filters.values()):
+                if not inside_geobounds(cha, **geo_filters):
+                    continue
 
             channels.append(cha)
         sta = copy.copy(self)
