### 2020-01-24 689d851
stream.select: make sure inventory restrictions still apply other restirctions as well
