### 2020-01-29 441ae6a
console_scripts: avoid hickup on Python 2
see
https://github.com/conda-forge/obspy-feedstock/pull/31#issuecomment-579767081

Just work around some weird erroneous SyntaxError check due to a module
being named "print.py", unfortunately
