+++ setup.py
@@ -131,7 +131,7 @@ ENTRY_POINTS = {
         'obspy-flinn-engdahl = obspy.scripts.flinnengdahl:main',
         'obspy-runtests = obspy.scripts.runtests:main',
         'obspy-reftek-rescue = obspy.scripts.reftekrescue:main',
-        'obspy-print = obspy.scripts.print:main',
+        'obspy-print = obspy.scripts._print:main',
         'obspy-sds-report = obspy.scripts.sds_html_report:main',
         'obspy-indexer = obspy.db.scripts.indexer:main',
         'obspy-scan = obspy.imaging.scripts.scan:main',
