+++ obspy/scripts/_print.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+Print stream information for waveform data in local files.
+"""
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from .print import main
+
+main = main
