### 2020-01-30 #16 @raoulcollenteur
I could help with a review. Not specifically known in seismology but I am a hydrogeologist and (co-) develop some python packages.  Let me know if help is needed :)
