+++ obspy/scripts/_print.py
@@ -5,6 +5,7 @@ Print stream information for waveform data in local files.
 """
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
+from future.builtins import *  # NOQA
 
 from .print import main
 
