### 2020-01-30 e8c0c38
shorten traceback by converting keysview to list
