### 2020-01-30 32c1de8
shorten line (pep8)
