+++ obspy/taup/tau_branch.py
@@ -8,8 +8,6 @@ from __future__ import (absolute_import, division, print_function,
 from future.builtins import *  # NOQA
 from future.utils import native_str
 
-import warnings
-
 import numpy as np
 
 from .c_wrappers import clibtau
