### 2020-01-30 6b90f8f
revert changes from #2280 to see potential error messages on CI
