### 2020-01-30 b23d30a
refactor _robust_resize, so that it does not show warning automatically
