### 2020-01-30 #16 @lwasser
hi @canyon289 and @raoulcollenteur  if you have experience developing packages, and are on the geosciences end of things, I think that could work well for this review. @megies please let me know what you think about two people reviewing who are not from the seismology community explicitly? i suspect this might be challenging to find someone without a conflict of interest and such.  I know both are technically qualified based upon the comments above!! Thank you all for your responses.
