### 2020-01-30 855cce0
update _robust_resize docstring
