### 2020-01-31 #16 @megies
> @megies please let me know what you think about two people reviewing who are not from the seismology community explicitly?

I don't see a problem with that at all. :+1:
