### 2020-01-31 57a9aa1
Fixes Issue #2543 (ObsPy writes SAC binary character header variables in the wrong format)
