+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -14,22 +14,22 @@ https://github.com/obspy/obspy/wiki/ObsPy-Git-Branching-Model
 
 ### What does this PR do?
 
-*Please fill in*
+Fixes issue [#2543](https://github.com/obspy/obspy/issues/2543#issuecomment-580663921)
 
 ### Why was it initiated?  Any relevant Issues?
 
-*Please fill in*
+Invited to by **megies** to include in 1.2.0 RC.
 
 ### PR Checklist
-- [ ] Correct base branch selected? `master` for new features, `maintenance_...` for bug fixes
+- [x] Correct base branch selected? `master` for new features, `maintenance_...` for bug fixes
 - [ ] This PR is not directly related to an existing issue (which has no PR yet).
 - [ ] If the PR is making changes to documentation, docs pages can be built automatically.
       Just remove the space in the following string after the + sign: "+ DOCS"
 - [ ] If any network modules should be tested for the PR, add them as a comma separated list
       (e.g. `clients.fdsn,clients.arclink`) after the colon in the following magic string: "+TESTS:"
       (you can also add "ALL" to just simply run all tests across all modules)
-- [ ] All tests still pass.
-- [ ] Any new features or fixed regressions are be covered via new tests.
+- [x] All tests still pass.
+- [x] Any new features or fixed regressions are be covered via new tests.
 - [ ] Any new or changed features have are fully documented.
-- [ ] Significant changes have been added to `CHANGELOG.txt` .
-- [ ] First time contributors have added your name to `CONTRIBUTORS.txt` .
+- [x] Significant changes have been added to `CHANGELOG.txt` .
+- [x] First time contributors have added your name to `CONTRIBUTORS.txt` .
