### 2020-01-31 #16 @lwasser
awesome @raoulcollenteur  and @canyon289 if you are still willing, I am going to assign you as reviewers for obspy!! Please plan to have your feedback to @megies by **Monday Feb 24th**.

You can use our review guide as a basis for your review. https://www.pyopensci.org/dev_guide/peer_review/reviewer_guide.html We actually have some reviews that you can look out now (our guide is dated already!!) which include:

* [earthpy:](https://github.com/pyOpenSci/software-review/issues/3)
* [pandera: ](https://github.com/pyOpenSci/software-review/issues/12)

@megies did check they are ok with issues being submitted directly to the obspy repo! so if you'd like you can organize that review with issues providing commentary here and then linking to issues that you open in obspy. how you prefer to work is up to you!

If you have any questions please get in touch here!! Thank you again for volunteering your time to review this package! It is a very widely used package in this domain  looking at the package download stats on conda so we are excited to see it submitted here!  Have a wonderful weekend!!
