+++ obspy/io/sac/tests/test_sactrace.py
@@ -451,14 +451,15 @@ class SACTraceTestCase(unittest.TestCase):
         )
         with NamedTemporaryFile() as tf:
             tempfile = tf.name
-        sac.write(tempfile,byteorder='big')
+        sac.write(tempfile, byteorder='big')
 
         with open(tempfile, "rb") as fh:
             str = fh.read()
         # kstnm is at offset 0x1b8 in the file
-        self.assertEqual(str[0x1b8:(0x1b8+8)],'TEST    ')
+        self.assertEqual(str[0x1b8:(0x1b8+8)], 'TEST    ')
         # kcmpnm is at offset 0x258 in the file
-        self.assertEqual(str[0x258:(0x258+8)],'Z       ')
+        self.assertEqual(str[0x258:(0x258+8)], 'Z       ')
+
 
 def suite():
     return unittest.makeSuite(SACTraceTestCase, 'test')
