### 2020-02-01 2ad03fd
Explicit bytes type-cast for Python 3.x
