### 2020-02-01 #16 @megies
> please excse me if i already asked you this? are you interested in a JOSS citation? it's not a significant amount of extra work

No problem. I'm not sure if it would do much though, since we have three citable papers for obspy but if you have a good idea how to theme it or if you think it's still adding something, let me know. :-)
