### 2020-02-01 #16 @megies
> Also interested in whether this is getting submitted to JOSS as there seem to be extra reviewer items for me to go through if that's the case

You can start on those items not affected by JOSS submission :-)
