### 2020-02-10 1893c60
Fix StrinIO-obj looping to properly read wavname
