### 2020-02-10 5170633
flake8 formatting fix
