### 2020-02-11 fc8e2f9
Call _readwavename with tagged_lines, not with StringIO-obj
