### 2020-02-11 4201281
Hacky fix for reading SC3 ArclinkXML where publicID has slash instead of shebang
