+++ obspy/io/arclink/tests/test_inventory_xml.py
@@ -44,8 +44,8 @@ class ArclinkInventoryTestCase(unittest.TestCase):
         self.station_afc_path = os.path.join(self.data_dir, "station_afc.xml")
 
     def test_publicid_slash(self):
-        inv = read_inventory(os.path.join(self.data_dir, "public-id-slash.xml"))
-        self.assertEqual(len(inv[0][0][0].response.response_stages), 11)
+        v = read_inventory(os.path.join(self.data_dir, "public-id-slash.xml"))
+        self.assertEqual(len(v[0][0][0].response.response_stages), 11)
 
     def test_analogue_filter_chain(self):
         """
