### 2020-02-12 f04b25c
adding RSHAKE datasource to docstring
