### 2020-02-12 c00387f
adding Raspberry Shake endpoint to URL mappings
