+++ obspy/CONTRIBUTORS.txt
@@ -67,6 +67,7 @@ Michelini, Alberto
 Miller, Nathaniel C.
 Morgenstern, Bernhard
 Murray-Bergquist, Louisa
+Nesbitt, Ian M.
 Nof, Ran Novitsky
 Panning, Mark P.
 Rapagnani, Giovanni
