### 2020-02-13 8a07a36
tests: avoid a floating point accuracy fail at 15th significant digit
http://tests.obspy.org/108581/#1
