### 2020-02-13 fa24a20
io.iaspei: fix an indexing error
see http://tests.obspy.org/108581/#2 and others

not sure why this only raises an index error on that specific machine,
we might have some inconsistencies in `Catalog.__bool__`??
