### 2020-02-13 384c5fe
tsindex: show warning and skip unittests for sqlalchemy<1.0
see http://tests.obspy.org/108600/#1

e.g. centos 7 has ancient sqlalchemy
