### 2020-02-13 688c323
switch mapping to RASPISHAKE
