+++ obspy/clients/fdsn/header.py
@@ -58,7 +58,7 @@ URL_MAPPINGS = {
     "ODC": "http://www.orfeus-eu.org",
     "ORFEUS": "http://www.orfeus-eu.org",
     "RESIF": "http://ws.resif.fr",
-    "RSHAKE": "http://fdsnws.raspberryshakedata.com",
+    "RASPISHAKE": "http://fdsnws.raspberryshakedata.com",
     "SCEDC": "http://service.scedc.caltech.edu",
     "TEXNET": "http://rtserve.beg.utexas.edu",
     "USGS": "http://earthquake.usgs.gov",
