### 2020-02-13 500225e
RASPISHAKE mapping
