### 2020-02-13 6d3bb07
exclude RASPISHAKE from MassDownloader by default
