### 2020-02-14 4329d46
docker: update fedora base images to current supprted releases
