+++ misc/docker/base_images/fedora_30/Dockerfile
@@ -1,20 +1,23 @@
-FROM fedora:26
+FROM fedora:30
 
-MAINTAINER Lion Krischer
+MAINTAINER Tobias Megies
 
 # Can fail on occasion.
 RUN dnf -y upgrade || true
+# looks like all "python-" is just links to respective python3 packages
+# might have to adapt package names for next release
 RUN dnf install -y \
+    python-devel \
     gcc \
     libmseed \
     libmseed-devel \
+    m2crypto \
     numpy \
-    pyshp \
-    python-GeographicLib \
-    python-basemap \
     python-basemap-data \
+    python-cryptography \
     python-decorator \
     python-future \
+    python-jsonschema \
     python-lxml \
     python-matplotlib \
     python-mock \
@@ -22,6 +25,10 @@ RUN dnf install -y \
     python-requests \
     python-sqlalchemy \
     python-tornado \
+    python3-GeographicLib \
+    python3-basemap \
+    python3-pyproj \
+    python3-pyshp \
     redhat-rpm-config \
     scipy
 RUN pip install https://github.com/Damgaard/PyImgur/archive/9ebd8bed9b3d0ae2797950876f5c1e64a560f7d8.zip
