### 2020-02-14 2b8a5ea
docker: update opensuse images
