### 2020-02-14 83b0fe1
docker: add centos_8 image
