### 2020-02-14 cdfe146
fix doctest
