### 2020-02-14 b9c7045
minor doctest fix
