### 2020-02-14 23747d1
seedlink tests: switch station used in test
seems other one has no data anymore
