### 2020-02-15 #16 @canyon289
Hi,
I've reviewed the package and have two comments

For the author
1. Tests are failing in CI and some fail locally.  http://tests.obspy.org/109013/
2. I wasn't able to quickly find a usage tutorial of the "intro to obspy" sort in the README. The docs are quite extensive so I'm sure one exists somewhere but I wasn't able to find it. The README shows an output image, perhaps the code used to generate that can be shown.

For the editors
1. There aren't examples for *every* user facing function, but there are for most and API for this library is so expansive I can sympathize with not having examples for all functionality. Nonetheless if go by the wording it states all functions need user facing documentation. How should the review proceed here?

In depth comments and examples are in this google doc
https://docs.google.com/document/d/1kmH3b9_X_KZok1wXAwVUhgjQTtDQg5xwIcjB5oA_WOQ/edit?usp=sharing
