### 2020-02-16 ee9b44f
taup: fix C inner loop helper code, thanks @1sbeller
