### 2020-02-21 5c12b16
sort changelog and merge duplicate fields
