### 2020-02-21 2261ecc
start entries in changelog with upper case, align entries
