+++ CHANGELOG.txt
@@ -30,7 +30,7 @@ institutions/companies and grants (in alphabetical order):
 
 Changes:
  - obspy.core:
-   * inventory objects have been adapted to StationXML 1.1 for details on
+   * Inventory objects have been adapted to StationXML 1.1 for details on
      changes see #2510 and
      https://github.com/FDSN/StationXML/blob/master/Changes.md
    * Fixed import of custom plugins (see #2423)
@@ -47,12 +47,12 @@ Changes:
    * UTCDateTime objects can now accept hour, minute, second, and microsecond
      values greater than their normal limits by setting the strict keyword
      argument to False (see #2232).
-   * fixed UTCDateTime(..., julday=366) for non-leap years. This was returning
+   * Fixed UTCDateTime(..., julday=366) for non-leap years. This was returning
      January 1st of the next year in case of non-leap years being used. Now it
      properly raises an out-of-bounds ValueError (see #2369)
    * When reading StationXML/SC3ML, make sure to properly read empty string
      fields as empty strings instead of "None" (see #2519 and #2527)
-   * better ISO8601 detection for UTCDateTime objects and UTCDateTime(...,
+   * Better ISO8601 detection for UTCDateTime objects and UTCDateTime(...,
      iso8601=False) now completely disables ISO8601 handling (see #2447)
    * Added replace method to UTCDateTime class (see #2077).
    * Added remove method to Inventory class (see #2088).
@@ -103,47 +103,47 @@ Changes:
      have been deselected due to the default location priorities setting. This
      is a pure usability improvement as it has been confusing users
      (see #2159).
-   * make sure that streams fetched via FDSN are properly trimmed to user
+   * Make sure that streams fetched via FDSN are properly trimmed to user
      requested times if data center serves additional data around the start/end
      (see #1887, #2298)
-   * fix a problem that could spam subprocesses that were not closed in routed
+   * Fix a problem that could spam subprocesses that were not closed in routed
      clients (see #2342 and #2344)
-   * make it possible to use signed EIDA tokens and also skip token validation
+   * Make it possible to use signed EIDA tokens and also skip token validation
      completely (see #2297)
-   * adding a mapping for RASPISHAKE
+   * Adding a mapping for RASPISHAKE
  - obspy.clients.filesystem.tsindex:
    * Add new Client & Indexer modules based on IRIS time series index (see
      #2206)
  - obspy.clients.iris:
-   * results of distaz method are now returned as native floats (see #2499)
+   * Results of distaz method are now returned as native floats (see #2499)
  - obspy.clients.neic:
-   * properly use specified timeout value (see #2450)
+   * Properly use specified timeout value (see #2450)
  - obspy.clients.seedlink:
-   * add method "get_info()" to fetch information on what
+   * Add method "get_info()" to fetch information on what
      networks/stations/locations/channels are served by the seedlink server
      (see #2405)
    * "get_waveforms()" can now be used with '*' and '?' wildcards in any part
-     of requested SEED ID, i.e. network, station, location and channel (see
-     #2405)
+      of requested SEED ID, i.e. network, station, location and channel (see
+      #2405)
  - obspy.clients.seishub:
    * Properly handle fetching poles and zeros in presence of multiple metadata
      files for a given station (see #2411)
  - obspy.geodetics:
-   * new utility function `inside_geobounds()` to check whether an object is
+   * New utility function `inside_geobounds()` to check whether an object is
      inside a geographic bound (see #2515)
  - obspy.imaging:
[…]
