### 2020-02-23 #16 @raoulcollenteur
- [X] **A statement of need** is provided.
- [X] **Installation instructions:** Good documentation for different forms of installation is provided. I tried an install on MacOS 10.15.1 and anaconda with Python 3.7 and that worked fine. I think support for P3.7 could be added?
- [X] **Function Documentation:** for all user-facing functions. API documentation is provided.
- [ ] **Examples** for all user-facing functions
- [X] **Community guidelines** a contributors guide is available, and looking at the increasing number of external contributors it is working.
- [X] **Metadata** A general email is provided, but everything is traceable through GitHub anyway.

Readme requirements
- [X] The package name is included (nice logo:))
- [X] Badges for continuous integration and test coverage, **the badge for pyOpenSci peer-review has not been included yet.**
- [X] Short description of goals of package with links is provided.
- [X] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage
- [X] Direction to more detailed documentation (e.g. your documentation files or website).

- [X] **Installation:** Installation succeeds as documented. Running the tests gave me two minor errors, and ObsPy gives the handy functionality to automatically update the package maintainers on this. Well done.
- [X] **Performance:** I could not find real claims on performance, so this is fine.
- [X] **Automated tests:**Running the tests gave me two minor errors, and ObsPy gives the handy functionality to automatically update the package maintainers on this.
- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

---

I am not a seismologist by training so I will limit to software technical/UI review. ObsPy is a Python package targeted at seismologists and seems to have a large and active user base. I think this is better proof for the need and usability of this package than anything else. I congratulate the authors on what seems to me a well accepted Python package. The Wiki and documentation are extensive. Some of the information feels a little dispersed, maybe these two sources could be merge to one website in the future to improve easy access to all these resources. The gallery gives a nice quick overview of the possibilities, including the necessary code. In addition to all these resources there are multiple articles introducing the software that provide a quick introduction. I also found a couple of useful repositories in the organisation account (e.g., https://github.com/obspy/docs) that include notebooks, these could be promoted from the main repository as well.

Major comments
- One thing that I think should be addressed before publishing the package is the succes of CI tests. The CI is failing for both MacOS/Linus and Windows (on 2020/02/23).
- I would suggest to create a single Jupyter Notebook that shows the most important functionality and runs on both local machines and Binder as an easy introduction to the package. I found it a little difficult to find a notebook to test the most important functionality (I now copied code from https://docs.obspy.org/tutorial/ into a notebook..)

Minor comments
- Add explicit link to documentation website (http://docs.obspy.org) on the readme so it's easier to find. Possibly replacing the "read more" link as it is not clear this link brings you.
- The wiki has a short three-line section describing how to get started. I suggest adding a short section "Getting started to the readme. It provides some good pointers on how to quickly check out what ObsPy can do (e.g., the gallery). This could replace the "usage demonstration" which I do not think makes sense to add to the readme for this package.
- Possibly start supporting python 3.7? I tested on 3.7 and it all seems to work well.
- I see that some commits are now pushed directly to the master branch, so that the master branch is not necessarily the stable version anymore (e.g., on 2020/02/23 the build is failing). It could be an idea to create an additional development branch where direct pushed are possible, and keep the master branch as the stable version (tagged by a GitHub release).
