### 2020-02-25 6d079d3
try to fix loading cdll on windows
seems on Windows we suddenly have issues with '../' in the path?
