+++ obspy/core/util/libnames.py
@@ -92,11 +92,22 @@ def _load_cdll(name):
     libpath = os.path.join(libdir, libname)
     # resolve parent directory '../' for windows
     libpath = os.path.normpath(libpath)
+    libpath = str(libpath)
     try:
-        cdll = ctypes.CDLL(str(libpath))
+        cdll = ctypes.CDLL(libpath)
     except Exception as e:
-        msg = 'Could not load shared library "%s" ("%s").\n\n %s' % (
-            libname, str(libpath), str(e))
+        import textwrap
+        dirlisting = textwrap.wrap(
+            ', '.join(sorted(os.path.dirname(libpath))))
+        msg = ['Could not load shared library "%s"' % libname,
+               'Path: %s' % libpath,
+               'Current directory: %s' % os.path.abspath(os.curdir),
+               'ctypes error message: %s' % str(e),
+               'Directory listing of lib directory:',
+               '  ',
+               ]
+        msg = '\n  '.join(msg)
+        msg = msg + '\n    '.join(dirlisting)
         raise ImportError(msg)
     return cdll
 
