### 2020-02-25 54e89dc
win DLL load: use abspath
