### 2020-02-26 c56db66
CI: deactivate 32bit builds on Windows/Appveyor
started having problems getting installed through conda
