### 2020-02-26 784da91
Appveyor command priority
