+++ appveyor.yml
@@ -85,11 +85,11 @@ install:
   # Check that we have the expected version and architecture for Python
   - "python -c \"import sys; print(sys.version)\""
 
-  # Install the build and runtime dependencies of the project.
-  - "conda update -q --yes conda"
   - "conda config --add channels conda-forge"
   # add channel twice to move it to the top of the channel priority list
   - "conda config --add channels conda-forge"
+  # Install the build and runtime dependencies of the project.
+  - "conda update -q --yes conda"
   # Create a conda environment using the required ObsPy packages.
   - "conda create -q --yes -n test python=%PYTHON_VERSION%"
   - "activate test"
