### 2020-02-26 fd75c1f
smaller fix appveyor and building on tom's machine
