+++ appveyor.yml
@@ -1,6 +1,9 @@
 # AppVeyor.com is a Continuous Integration service to build and run tests under Windows
 # see https://github.com/ogrisel/python-appveyor-demo/blob/master/appveyor.yml
 
+build_cloud: PC1577-AS
+image: Windows
+
 clone_depth: 50
 
 environment:
@@ -97,10 +100,11 @@ install:
   # pip 10.0.1 seems buggy and breaks all our appveyor builds
   - ps: |
       If ($env:PYTHON_VERSION -eq "3.8") {
-          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests jsonschema cryptography proj4
+        SET BASEMAP=""
       }else {
-          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests basemap jsonschema pyshp cryptography proj4
+        SET BASEMAP=basemap
       }
+  - conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests %BASEMAP% jsonschema pyshp cryptography proj4
   # somehow basemap/pyproj/proj4 can fail when PROJ_LIB is not set correctly
   # for whatever reason, see:
   #  - conda-forge/basemap-feedstock#36
