### 2020-02-26 fa33a72
signal: fix covariance matrix, make adative windowing optional in vidale
patch by @jwassermann
