+++ obspy/signal/polarization.py
@@ -192,7 +192,7 @@ def instantaneous_frequency(data, sampling_rate):
     return instf
 
 
-def vidale_adapt(stream, noise_thres, fs, flow, fhigh, spoint, stime, etime):
+def vidale_adapt(stream, noise_thres, fs, flow, fhigh, spoint, stime, etime,adaptive=True):
     """
     Adaptive window polarization analysis after [Vidale1986]_ with the
     modification of adapted analysis window estimated by estimating the
@@ -217,6 +217,9 @@ def vidale_adapt(stream, noise_thres, fs, flow, fhigh, spoint, stime, etime):
     :type stime: :class:`~obspy.core.utcdatetime.UTCDateTime`
     :param etime: end time for the analysis
     :type etime: :class:`~obspy.core.utcdatetime.UTCDateTime`
+    :param adaptive: switch for adaptive window estimation defaults to True
+        if set to False the window will be estimated as 3*max(1/(fhigh-flow),1/flow)
+    :type adaptive: bool
     :returns: list of tuples containing azimuth, incidence, rectilinearity,
         planarity, and ellipticity
     """
@@ -238,15 +241,18 @@ def vidale_adapt(stream, noise_thres, fs, flow, fhigh, spoint, stime, etime):
     offset = int(3 * fs / flow)
     covmat = np.zeros([3, 3], dtype=np.complex128)
     while True:
-        adapt = int(3. * w * fs / (zi[offset] + ni[offset] + ei[offset]))
+        if adaptive == True:
+            adapt = int(3. * w * fs / (zi[offset] + ni[offset] + ei[offset]))
         # in order to account for errors in the inst freq estimation
-        if adapt > int(3 * fs / flow):
-            adapt = int(3 * fs / flow)
-        if adapt < int(3 * fs / fhigh):
-            adapt = int(3 * fs / fhigh)
-        # XXX: was adapt /= 2
-        adapt //= 2
-        adapt = (2 * adapt) + 1
+            if adapt > int(3 * fs / flow):
+                adapt = int(3 * fs / flow)
+            if adapt < int(3 * fs / fhigh):
+                adapt = int(3 * fs / fhigh)
+            # XXX: was adapt /= 2
+            adapt //= 2
+            adapt = (2 * adapt) + 1
+        else:
+            adapt = np.maximum(int(3. * fs/(fhigh-flow )),int(3. * fs/flow))
         newstart = stime + offset / fs
         if (newstart + (adapt / 2) / fs) > etime:
             break
@@ -261,16 +267,16 @@ def vidale_adapt(stream, noise_thres, fs, flow, fhigh, spoint, stime, etime):
         nx -= nx.mean()
         ex -= ex.mean()
 
-        covmat[0][0] = np.dot(ex, ex.conjugate())
-        covmat[0][1] = np.dot(ex, nx.conjugate())
-        covmat[1][0] = covmat[0][1].conjugate()
-        covmat[0][2] = np.dot(ex, zx.conjugate())
-        covmat[2][0] = covmat[0][2].conjugate()
-        covmat[1][1] = np.dot(nx, nx.conjugate())
-        covmat[1][2] = np.dot(zx, nx.conjugate())
-        covmat[2][1] = covmat[1][2].conjugate()
-        covmat[2][2] = np.dot(zx, zx.conjugate())
+        mask = (stream[0][:] ** 2 + stream[1][:] ** 2 + stream[2][:] ** 2) > noise_thres
+        xx = np.zeros((3, mask.sum()), dtype=np.complex128)
+        # East
+        xx[0, :] = ea
+        # North
+        xx[1, :] = na
+        # Z
+        xx[2, :] = za
 
+        covmat = np.cov(xx)
[…]
