### 2020-02-26 cd3bd61
flake8 and changelog
