### 2020-02-26 aceb0f3
smaller fix appveyor and building on tom's machine
