+++ appveyor.yml
@@ -83,7 +83,7 @@ install:
   # Prepend Python to the PATH of this build (this cannot be
   # done from inside the powershell script as it would require to restart
   # the parent CMD process).
-  - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
+  - "SET PATH=D:\BUILD\miniconda3;D:\BUILD\miniconda3\Scripts;%PATH%"
 
   # Check that we have the expected version and architecture for Python
   - "python -c \"import sys; print(sys.version)\""
