### 2020-02-26 eef39bf
smaller fix appveyor and building on tom's machine
