### 2020-02-26 2146c9a
tiny code snippet and image output for README
