### 2020-02-26 95e2495
add note on seismo-live Tutorial resources in README
