### 2020-02-26 #16 @megies
Here's my reply to @canyon289, thanks for the helpful comments. I also left some remarks in the google docs, mainly about those failing tests in CI.

### 1

We quite often have tests failing which most times is due to the tests being so rigorous. I can make you a log of numpy floating point operation result differences at the 15th digit and their changes across platforms, architectures and numpy versions. Fun aside, all of the fails you linked are image test fails, you can actually check out the images uploaded to imgur from the traceback message. Usually it's ticks moving around a bit.

### 2

Usage examples could be found following the `read more` link to the wiki and move to `Getting started` section referring to the Tutorial in the docs. I guess it was not perfectly visible, so I made the `read more` stand out a bit more and added a small `Getting Started` section as well.
The plot you were referring to is not really an ObsPy plot, but rather plotting our usage stats, so it's not appropriate to show the code for it, but I added a mini code example with a picture as a teaser in the README. Thanks for the suggestion to make usage teaser stand out a bit more in the README.

```python
from obspy import read
st = read()  # load example seismogram
st.filter(type='highpass', freq=3.0)
st = st.select(component='Z')
st.plot()
```

![plot](https://user-images.githubusercontent.com/1842780/75336536-a0d16f80-588b-11ea-812c-7b20b6848865.png)
