### 2020-02-26 #16 @megies
@raoulcollenteur, thanks for your review, here's my comments/replies:

 - @canyon289 also commented about the missing usage example in README, I added it to the README and also improved the "Getting started" in the wiki. I also added mention of www.seismo-live.org which has static preview notebooks as well as binder for interactive use of all the ObsPy training notebooks (and much more).
 - I don't really understand the `Vignette(s) demonstrating major functionality that runs successfully locally` item actually @lwasser. In any case all the important submodule API landing pages have extensive examples (in addition to examples in specific important method/function API pages), so I'm pretty confident we're good there, e.g. http://docs.obspy.org/master/packages/obspy.clients.fdsn.html
 - again, about the tests, same comments like to @canyon289 apply. many test fails are because tests are a) so rigorous, b) new matplotlib version decides to place ticks 2 pixels further left, c) network/server side issues we don't have control over. I also hate not see green all the time, but we tend to accept to have some more fails sometimes but be sure that we can actually see it when something really breaks
 - [notebooks with the most basic usage are in place](https://krischer.github.io/seismo_live_build/html/ObsPy/00_Introduction_wrapper.html), also mentioned in the top of the "Tutorial" docs page. All of that is also on binder via seismo-live, like stated above I added better mention and links to that in README and wiki "Getting started".
 - `Possibly start supporting python 3.7?` We do actually, and have 3.7 and 3.8 also in CI, I guess you are referring to what is stated on PyPi? Those metadata will get updated by uploading the next release, [its already fixed in master](https://github.com/obspy/obspy/blob/d306560b8f23d2a03939d949565f65f592bf3428/setup.py#L802-L810).
 - `I see that some commits are now pushed directly to the master branch [...]`  It's an interesting idea but it might confuse some of our less git/github-savvy contributors even more. It already is confusing for some to have a master branch and then a maintenance branch for fixes only.
