### 2020-02-26 1d857c3
travis: pin pip on py3.4 (will drop it afterwar release though)
