+++ .travis.yml
@@ -95,6 +95,7 @@ install:
         BASEMAP="basemap=1.0.7"
         PYPROJ="pyproj"
         GFORTRAN="libgfortran=1"
+        PIP="pip"
         PROJ4="proj4"
         # ancient matplotlib needs to be turned to AGG before anything else,
         # otherwise it tries to import incompatible version of Qt as a backend
@@ -104,10 +105,12 @@ install:
       elif [[ "${PYTHON_VERSION}" == '3.4' ]]; then
         # 3.4 apparently no longer get's updated packages.
         # Let conda resolve to the latest versions.
+        # pip dropped py3.4 support with 12.1
         NUMPY="numpy"
         SCIPY="scipy"
         MATPLOTLIB="matplotlib"
         BASEMAP="basemap"
+        PIP="pip<19.2"
         PYPROJ="pyproj"
         PROJ4="proj4"
         GFORTRAN=""
@@ -127,6 +130,7 @@ install:
         SCIPY="scipy"
         MATPLOTLIB="matplotlib"
         BASEMAP="basemap"
+        PIP="pip"
         PYPROJ="pyproj"
         PROJ4="proj4=4.9.3"
         GFORTRAN=""
@@ -137,6 +141,7 @@ install:
         SCIPY="scipy"
         MATPLOTLIB="matplotlib"
         BASEMAP="basemap>=1.2.1"
+        PIP="pip"
         PYPROJ="pyproj"
         PROJ4="proj4"
         GFORTRAN=""
@@ -145,6 +150,7 @@ install:
         SCIPY="scipy"
         MATPLOTLIB="matplotlib>=2.0.0"
         BASEMAP="basemap"
+        PIP="pip"
         PYPROJ="pyproj"
         PROJ4="proj4"
         GFORTRAN=""
@@ -156,6 +162,7 @@ install:
         $SCIPY
         $MATPLOTLIB
         $BASEMAP
+        $PIP
         $PYPROJ
         $PROJ4
         $GFORTRAN
