### 2020-02-26 ac48350
travis: install codecov from conda, packaged now
eases some pip install problems in CI when pip wants to uninstall
conda-installed packaged for downgrading etc
