+++ appveyor.yml
@@ -33,22 +33,6 @@ environment:
       PYTHON_VERSION: "3.8"
       PYTHON_ARCH: "64"
 
-
-init:
-  # If there is a newer build queued for the same PR, cancel this one.
-  # The AppVeyor 'rollout builds' option is supposed to serve the same
-  # purpose but it is problematic because it tends to cancel builds pushed
-  # directly to master instead of just PR builds (or the converse).
-  # credits: JuliaLang developers.
-  - ps: if ($env:APPVEYOR_PULL_REQUEST_NUMBER -and $env:APPVEYOR_BUILD_NUMBER -ne ((Invoke-RestMethod `
-        https://ci.appveyor.com/api/projects/$env:APPVEYOR_ACCOUNT_NAME/$env:APPVEYOR_PROJECT_SLUG/history?recordsNumber=50).builds | `
-        Where-Object pullRequestId -eq $env:APPVEYOR_PULL_REQUEST_NUMBER)[0].buildNumber) { `
-          throw "There are newer queued builds for this pull request, failing early." }
-  - ps: |
-      If (($env:SKIP_NOTAG -eq "true") -and ($env:APPVEYOR_REPO_TAG -ne "true")) {
-          throw "Skipping Python version, not a tag."
-      }
-
 install:
   # - ECHO "Filesystem root:"
   # - ps: "ls \"C:/\""
@@ -60,6 +44,7 @@ install:
   # done from inside the powershell script as it would require to restart
   # the parent CMD process).
   - "SET PATH=D:\\BUILD\\miniconda3;D:\\BUILD\\miniconda3\\Scripts;%PATH%"
+  - D:\BUILD\miniconda3\Scripts\activate.bat D:\BUILD\miniconda3
   # Check that we have the expected version and architecture for Python
   - "python -c \"import sys; print(sys.version)\""
   - "conda config --add channels conda-forge"
