### 2020-02-26 430bc8f
delete envs when finished
