### 2020-02-26 5e70eba
travis: try to fix py3.5 build (pip install trying to upgrade certifi)
