+++ appveyor.yml
@@ -25,6 +25,15 @@ environment:
     # a later point release.
     # See: http://www.appveyor.com/docs/installed-software#python
 
+    - PYTHON: "C:\\Miniconda35-x64"
+      PYTHON_VERSION: "3.5"
+      PYTHON_ARCH: "64"
+      SKIP_NOTAG: "true"
+
+    - PYTHON: "C:\\Miniconda36-x64"
+      PYTHON_VERSION: "3.6"
+      PYTHON_ARCH: "64"
+
     - PYTHON: "C:\\Miniconda37-x64"
       PYTHON_VERSION: "3.7"
       PYTHON_ARCH: "64"
@@ -95,8 +104,10 @@ on_success:
           python setup.py sdist bdist_wheel bdist_wininst bdist_egg
           Invoke-Expression "twine upload -u $env:PYPI_USERNAME -p $env:PYPI_PASSWORD --repository-url https://upload.pypi.org/legacy/ --skip-existing dist/*" 2>&1 | Out-Null
       }
-  - "conda deactivate"
-  - "conda env remove -n %ENV_NAME%"
-
+  
 # Do not build feature branch with open Pull Requests
 skip_branch_with_pr: true
+
+on_finish:
+  - "conda deactivate"
+  - "conda env remove -n %ENV_NAME%"
