### 2020-02-26 2a89fd5
travis: unquiet conda install to see whats getting resolved and installed right away
