+++ appveyor.yml
@@ -25,20 +25,20 @@ environment:
     # a later point release.
     # See: http://www.appveyor.com/docs/installed-software#python
 
-    - PYTHON: "C:\\Miniconda35-x64"
+    - PYTHON: "D:\\BUILD\\miniconda3"
       PYTHON_VERSION: "3.5"
       PYTHON_ARCH: "64"
       SKIP_NOTAG: "true"
 
-    - PYTHON: "C:\\Miniconda36-x64"
+    - PYTHON: "D:\\BUILD\\miniconda3"
       PYTHON_VERSION: "3.6"
       PYTHON_ARCH: "64"
 
-    - PYTHON: "C:\\Miniconda37-x64"
+    - PYTHON: "D:\\BUILD\\miniconda3"
       PYTHON_VERSION: "3.7"
       PYTHON_ARCH: "64"
 
-    - PYTHON: "C:\\Miniconda37-x64"
+    - PYTHON: "D:\\BUILD\\miniconda3"
       PYTHON_VERSION: "3.8"
       PYTHON_ARCH: "64"
 
@@ -52,27 +52,27 @@ install:
   # Prepend Python to the PATH of this build (this cannot be
   # done from inside the powershell script as it would require to restart
   # the parent CMD process).
-  - "SET PATH=D:\\BUILD\\miniconda3;D:\\BUILD\\miniconda3\\Scripts;%PATH%"
-  - D:\BUILD\miniconda3\Scripts\activate.bat D:\BUILD\miniconda3
+  - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
+  # Activate the base environment
+  - %PYTHON%\\Scripts\\activate.bat %PYTHON%
   # Check that we have the expected version and architecture for Python
   - "python -c \"import sys; print(sys.version)\""
-  - "conda config --add channels conda-forge"
-  # add channel twice to move it to the top of the channel priority list
-  - "conda config --add channels conda-forge"
-  # Install the build and runtime dependencies of the project.
-  - "conda update -q --yes conda"
-  # Create a conda environment using the required ObsPy packages.
-  - "conda create -q --yes -n %ENV_NAME% python=%PYTHON_VERSION%"
-  - "conda activate %ENV_NAME%"
-  # Install default dependencies
-  # pip 10.0.1 seems buggy and breaks all our appveyor builds
   - ps: |
       If ($env:PYTHON_VERSION -eq "3.8") {
         SET BASEMAP=""
       }else {
         SET BASEMAP=basemap
       }
-  - conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests %BASEMAP% jsonschema pyshp cryptography proj4
+  # Create a conda environment using the required ObsPy packages.
+  # - "conda create -q --yes -n %ENV_NAME% python=%PYTHON_VERSION%"
+  # Clones the exiting env for this PYTHON_VERSION number
+  - conda create -n %ENV_NAME% --clone py%PYTHON_VERSION%
+  - "conda activate %ENV_NAME%"
+  # Install default dependencies
+  # pip 10.0.1 seems buggy and breaks all our appveyor builds
+
+  # - conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests %BASEMAP% jsonschema pyshp cryptography proj4 pyimgur future geographiclib
+  
   # somehow basemap/pyproj/proj4 can fail when PROJ_LIB is not set correctly
   # for whatever reason, see:
   #  - conda-forge/basemap-feedstock#36
@@ -80,11 +80,8 @@ install:
[…]
