+++ appveyor.yml
@@ -59,11 +59,11 @@ install:
   - "python -c \"import sys; print(sys.version)\""
   - ps: |
       If ($env:PYTHON_VERSION -eq "3.8") {
-        conda create -q --yes -n %ENV_NAME% python=%PYTHON_VERSION%
-        conda activate %ENV_NAME%
+        conda create -q --yes -n %ENV_NAME% python=%PYTHON_VERSION% &&
+        conda activate %ENV_NAME% &&
         conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests jsonschema pyshp cryptography proj4 pyimgur future geographiclib
       }else {
-        conda create -n %ENV_NAME% --clone py%PYTHON_VERSION%
+        conda create -n %ENV_NAME% --clone py%PYTHON_VERSION% &&
         conda activate %ENV_NAME%
       }
   # somehow basemap/pyproj/proj4 can fail when PROJ_LIB is not set correctly
