### 2020-02-27 593c34d
remove env dir at the end
