### 2020-02-27 a812c7e
offline mode for py<3.8
