+++ appveyor.yml
@@ -63,7 +63,7 @@ install:
         conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests jsonschema pyshp cryptography proj4 pyimgur future geographiclib
       )
     ELSE (
-        conda create -n %ENV_NAME% --clone py%PYTHON_VERSION% &&
+        conda create -n %ENV_NAME% --clone py%PYTHON_VERSION% --offline &&
         conda activate %ENV_NAME%
       )
   # somehow basemap/pyproj/proj4 can fail when PROJ_LIB is not set correctly
