### 2020-02-28 4d40064
Revert "docker: run all tests when packaging and testing debs"
This reverts commit e3dc948b2f75402e3612dcc3cf089d4b14914a65.

Just takes too long to run all network tests
