+++ misc/docker/scripts/package_debs.sh
@@ -19,8 +19,8 @@ else
     echo -e "${green}Deb packaging successful!${no_color}"
     touch /build.success
     dpkg -i /tmp/python-obspy_build/packages/python-obspy-dbg_*.deb 2>&1 | tee /TEST_LOG.txt
-    dpkg -i /tmp/python-obspy_build/packages/python-obspy_*.deb && apt-get install -f && ((obspy-runtests -r --all --keep-images --no-flake8 --node=docker-deb-$(cat /container_name.txt) 2>&1 | tee --append /TEST_LOG.txt) && touch /test.success || touch /test.failure)
-    dpkg -i /tmp/python-obspy_build/packages/python3-obspy_*.deb && apt-get install -f && ((obspy3-runtests -r --all --keep-images --no-flake8 --node=docker-deb-$(cat /container_name.txt) 2>&1 | tee --append /TEST_LOG.txt) && touch /test.success || touch /test.failure)
+    dpkg -i /tmp/python-obspy_build/packages/python-obspy_*.deb && apt-get install -f && ((obspy-runtests -r --keep-images --no-flake8 --node=docker-deb-$(cat /container_name.txt) 2>&1 | tee --append /TEST_LOG.txt) && touch /test.success || touch /test.failure)
+    dpkg -i /tmp/python-obspy_build/packages/python3-obspy_*.deb && apt-get install -f && ((obspy3-runtests -r --keep-images --no-flake8 --node=docker-deb-$(cat /container_name.txt) 2>&1 | tee --append /TEST_LOG.txt) && touch /test.success || touch /test.failure)
 fi
 
 echo "Done with everything!"
