### 2020-02-28 a680bce
deb: try to avoid some weird problems while switching branches
