### 2020-02-28 4e7cf6d
compatibility.py: be nice to some older Linux distros with Python 3.4 like Debian jessie LTS
