### 2020-02-28 fdab3d4
deb: try to fix deb build on jessie
