### 2020-02-28 #16 @lwasser
thank you @megies !! i'll look at this a bit more closely today!
one note - when we talk about vignettes we like to see them in the api docs but also something like this is always nice:

* [earthpy vignette gallery](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html)
* [geopandas vignette gallery](https://geopandas.readthedocs.io/en/latest/gallery/index.html)

It's becoming clear to me that we need to more carefully consider levels of documentation for good, better and best . The issue of API documentation with examples for instance has come up multiple times. I'm curious what your take is on this. We have an discussion on it in discourse started here. @canyon289  and @raoulcollenteur  i'm very interested in your thoughts as well if you have any input! All - thank you for supporting / going through review process and for your patience as we develop our policies and guidelines!
