### 2020-02-28 #16 @lwasser
also @canyon289  and @raoulcollenteur  can you kindly have a look at the comments above and let me know if you are satisfied with @megies  responses? i see that you guys did go back and forth in the google doc! i will copy that text over to this issue just so we have a record of it now (in case the doc gets lost in the future!).
