### 2020-02-28 #16 @lwasser
### THIS IS THE GOOGLE DOC REVIEW TEMPLATE - from @canyon289 copying it here for records and safe keeping!! ###
Review Template

Checklist Legend
X - Complete and meets criteria
? - Not sure if criteria is met
Not Met - Criteria is not met
Not applicable - Not required for this review

No conflict of interest!

Short description is available on README. Link to a very comprehensive presentation is linked on README to youtube.

Installation instructions for dev is linked in README
https://github.com/obspy/obspy/wiki#installation

A gallery exists in the docs http://docs.obspy.org/gallery.html

- [x] **Function Documentation:** for all user-facing functions
Documentation is available here. API is expansive, which is a good thing, so I didn’t verify if “all” user facing functions was documented but this looks really comprehensive.

- [?] **Examples** for all user-facing functions
There aren’t example for all user facing functions. For example
http://docs.obspy.org/packages/autogen/obspy.geodetics.base.calc_vincenty_inverse.html#obspy.geodetics.base.calc_vincenty_inverse
http://docs.obspy.org/packages/obspy.clients.neic.html#module-obspy.clients.neic

But there are a number that are very well documented
http://docs.obspy.org/packages/obspy.core.html#module-obspy.core
How strict is the word all here?

Code of Conduct is available and is very much in line with inclusive community standard

- [X] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.
Contributors list avaiable in repo and gitter is available for questions. I believe this enompasses the metadata question
https://github.com/obspy/obspy/blob/master/obspy/CONTRIBUTORS.txt

Readme requirements

- [X] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.

Badge for review has not been included yet but everything else has

- [X] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [X] Any additional setup required (authentication tokens, etc)

- [Not Met] Brief demonstration usage
A brief demonstration of usage is not included at this time

- [X] Direction to more detailed documentation (e.g. your documentation files or website).

- [?] If applicable, how the package compares to other similar packages and/or how it relates to other packages
Unaware of what similar packages would be in this space

Installation from conda forge was well documented and painless

I’m not an expert enough with seismology to deeply evaluate features. From documentation it looks like what is promised does work

- [Not Applicable] **Performance:** Any performance claims of the software been confirmed.
I assume it works. I’m not qualified enough in seismology to deeply verify this one

- [Fail] **Automated tests:** Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

Tests do not pass on my local machine with fresh installation from master (commit: 34c70be39b71c6ec224a38045bb2a89a75f56a76)

When pyimgur is installed tests then fail because they’re hitting a web api that is sending back 400.

http://tests.obspy.org/109013/

This test failure service you’ve implemented though is really cool.

- [?] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

CI integration exists for appveyor and Travis but Travis seems to be failing at the moment

#### For packages co-submitting to JOSS

- [Not Applicable] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements).

- [Not Applicable] **A short summary** describing the high-level functionality of the software
- [Not Applicable] **Authors:** A list of authors with their affiliations
- [Not Applicable] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
- [Not Applicable] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

### Fix: Develop installation
Developer installation isn’t straightforward. For editable mode. It seems that numpy is needed prior to using pip, but this is not listed in the instructions
https://github.com/obspy/obspy/wiki/Developer-Installation
### Suggestion: Intro for dummies
A brief abstract about how this package is useful for seismologists would be great for non seismologists like me. There is an hour youtube video available but that doesn’t work for someone scrolling through
