### 2020-03-03 #16 @raoulcollenteur
Thanks for the comments on the review @megies, I think all the questions were answered well. Just the last thing I am not really sure about is the failing of CI tests (Travis fails: https://travis-ci.org/obspy/obspy). I agree with @canyon289 that this is rather confusing and could be fixed by allowing some tests to fail. I think many people use this to quickly evaluate a package for a) practical use / quality and b) ensure functioning during development of new features. In general, I would like to see the CI tests work for accepted packages in pyOpenSci.

> * @canyon289 also commented about the missing usage example in README, I added it to the README and also improved the "Getting started" in the wiki. I also added mention of [www.seismo-live.org](http://www.seismo-live.org) which has static preview notebooks as well as binder for interactive use of all the ObsPy training notebooks (and much more).

The links to seismic live don't work for me, you might want to check those. Readme looks good now.

> * I don't really understand the `Vignette(s) demonstrating major functionality that runs successfully locally` item actually @lwasser. In any case all the important submodule API landing pages have extensive examples (in addition to examples in specific important method/function API pages), so I'm pretty confident we're good there, e.g. http://docs.obspy.org/master/packages/obspy.clients.fdsn.html

I agree with @megies, I think this is fine as is. @lwasser I feel PyOpenSci should leave the exact implementation to the developers and not mandate that these vignettes should be in the API. It is nice, but for some applications it is not feasible in my opinion.

> * `Possibly start supporting python 3.7?` We do actually, and have 3.7 and 3.8 also in CI, I guess you are referring to what is stated on PyPi? Those metadata will get updated by uploading the next release, [its already fixed in master](https://github.com/obspy/obspy/blob/d306560b8f23d2a03939d949565f65f592bf3428/setup.py#L802-L810).

I was indeed talking about the Pypi badges, great!
