### 2020-03-04 43f0edf
docstring fix (no SEG2 write functionality)
fixes #2572
