### 2020-03-05 #16 @megies
@lwasser no worries and I hope everything got sorted out :+1:

    **good enough (required)**: all user facing functions (inputs and outputs) are documented both in the code and in the user-facing documentation.
    **better**: all user facing functions are documented with examples (in tutorials or doc strings)
    **best**: all user facing functions are documented and there is are tutorials (vignettes) that demonstrate example user workflows

@lwasser This looks like a good starting point for a classification. From my experience it might be enough to say (for e.g. "better") `"all user facing functions are documented in docstrings, most important or likely-to-be-used-often functions include examples in doc strings and/or tutorials"`.

But then I have to be self-critical here, obspy has grown from small to very big and some policies for better coding practice were picked up one by one. So at some point we set a policy that all functions/methods/classes deemed private are prefixed by one underscore, and subject to change of API or complete removal without warning or deprecation at devs discretion. For these I would also not demand full documentation or examples in most cases. But, we have lots of such functionality that is not marked private by naming, but we would still consider it more internal helpers that do not need full effort in docstrings or even examples. So I guess in our case you could say that it is not straight forward to tell what is really supposed to be deemed "user-facing". These are historical artifacts and may be similar (or not..) for other projects that were not laid out on the drafting table but rather have grown steadily.

Also, I think many smaller, yet user-facing routines might not not full blown tutorials. As an example, this helper function is for users, but I would argue the docstring says it all: http://docs.obspy.org/master/packages/autogen/obspy.geodetics.base.gps2dist_azimuth.html
