### 2020-03-05 137c2f9
fix a doctest by skipping non interesting parts
