### 2020-03-05 33887c9
testing: use BytesIO over a temp file in a SAC test
