+++ CHANGELOG.txt
@@ -1,4 +1,4 @@
-1.2.0rc8
+1.2.0rc9
 ========
 
 Work on this release was in parts and among others supported by the following
