### 2020-03-06 4bdbb67
release ObsPy 1.2.0
