### 2020-03-09 0b6223b
sunset py2: remove Py2 from setup.py and CI
