+++ obspy/io/segy/header.py
@@ -3,10 +3,6 @@
 Defines the header structures and some other dictionaries needed for SEG Y read
 and write support.
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
 import numpy as np
 
 from . import pack, unpack
