+++ obspy/io/seg2/__init__.py
@@ -13,11 +13,6 @@ format.
     (https://www.gnu.org/copyleft/lesser.html)
 
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
-
 if __name__ == '__main__':
     import doctest
     doctest.testmod(exclude_empty=True)
