+++ obspy/core/tests/test_network.py
@@ -9,10 +9,6 @@ Test suite for the network class.
     GNU Lesser General Public License, Version 3
     (https://www.gnu.org/copyleft/lesser.html)
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
 import io
 import os
 import unittest
