+++ obspy/clients/httpproxy.py
@@ -10,12 +10,6 @@ License:
 J. MacCarthy, modified from https://gist.github.com/frxstrem/4487802
 
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-from future import standard_library
-from future.utils import native_str
-
 from base64 import b64encode
 import socket
 with standard_library.hooks():
