+++ obspy/scripts/_print.py
@@ -3,10 +3,6 @@
 """
 Print stream information for waveform data in local files.
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
 from .print import main
 
 main = main
