+++ obspy/clients/arclink/tests/test_decrypt.py
@@ -2,10 +2,6 @@
 """
 The obspy.clients.arclink.client test suite.
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
 import os
 import unittest
 
