+++ obspy/signal/spectral_estimation.py
@@ -17,11 +17,6 @@ Various Routines Related to Spectral Estimation
     GNU Lesser General Public License, Version 3
     (https://www.gnu.org/copyleft/lesser.html)
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-from future.utils import native_str
-
 import bisect
 import glob
 import math
