+++ obspy/geodetics/__init__.py
@@ -9,10 +9,6 @@ obspy.geodetics - Various geodetic utilities for ObsPy
     GNU Lesser General Public License, Version 3
     (https://www.gnu.org/copyleft/lesser.html)
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
 from .base import (calc_vincenty_inverse, degrees2kilometers,
                    gps2dist_azimuth, inside_geobounds,
                    kilometer2degrees, kilometers2degrees, locations2degrees)
