+++ obspy/io/sh/evt.py
@@ -9,10 +9,6 @@ SeismicHandler evt file bindings to ObsPy core module.
     GNU Lesser General Public License, Version 3
     (https://www.gnu.org/copyleft/lesser.html)
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
 from collections import defaultdict
 import io
 from math import cos, pi
