### 2020-03-09 8d2bbcc
sunset py2: remove test for future/__future__ import headers
