### 2020-03-09 32568a5
sunset py2: remove usage of future's native_str() to convert strings
