+++ obspy/io/mseed/tests/test_mseed_special_issues.py
@@ -186,7 +186,7 @@ class MSEEDSpecialIssueTestCase(unittest.TestCase):
             # read temp file directly without libmseed
             with open(tempfile, 'rb') as fp:
                 fp.seek(56)
-                dtype = np.dtype(native_str('>f4'))
+                dtype = np.dtype('>f4')
                 bin_data = from_buffer(fp.read(7 * dtype.itemsize),
                                        dtype=dtype)
             np.testing.assert_array_equal(data, bin_data)
@@ -414,7 +414,7 @@ class MSEEDSpecialIssueTestCase(unittest.TestCase):
             # 5 - transform to ASCII values
             st = read()
             for tr in st:
-                tr.data = tr.data.astype(native_str('|S1'))
+                tr.data = tr.data.astype('|S1')
             # write a single trace automatically detecting encoding
             st[0].write(tempfile, format="MSEED")
             # write a single trace automatically detecting encoding
