+++ obspy/io/dmx/core.py
@@ -10,31 +10,31 @@ from obspy.core.compatibility import from_buffer
 from obspy.core.util.attribdict import AttribDict
 
 
-descript_trace_dtypes = np.dtype([(native_str("network"), native_str("4S")),
-                                  (native_str("st_name"), native_str("5S")),
-                                  (native_str("component"), native_str("1S")),
-                                  (native_str("insstype"), np.int16),
-                                  (native_str("begintime"), np.double),
-                                  (native_str("localtime"), np.int16),
-                                  (native_str("datatype"), native_str("1S")),
-                                  (native_str("descriptor"), native_str("1S")),
-                                  (native_str("digi_by"), np.int16),
-                                  (native_str("processed"), np.int16),
-                                  (native_str("length"), np.int32),
-                                  (native_str("rate"), np.float32),
-                                  (native_str("mindata"), np.float32),
-                                  (native_str("maxdata"), np.float32),
-                                  (native_str("avenoise"), np.float32),
-                                  (native_str("numclip"), np.int32),
-                                  (native_str("timecorrect"), np.double),
-                                  (native_str("rate_correct"), np.float32),
+descript_trace_dtypes = np.dtype([("network", "4S"),
+                                  ("st_name", "5S"),
+                                  ("component", "1S"),
+                                  ("insstype", np.int16),
+                                  ("begintime", np.double),
+                                  ("localtime", np.int16),
+                                  ("datatype", "1S"),
+                                  ("descriptor", "1S"),
+                                  ("digi_by", np.int16),
+                                  ("processed", np.int16),
+                                  ("length", np.int32),
+                                  ("rate", np.float32),
+                                  ("mindata", np.float32),
+                                  ("maxdata", np.float32),
+                                  ("avenoise", np.float32),
+                                  ("numclip", np.int32),
+                                  ("timecorrect", np.double),
+                                  ("rate_correct", np.float32),
                                   ])
 
-structtag_dtypes = np.dtype([(native_str("sinc"), native_str("1S")),
-                             (native_str("machine"), native_str("1S")),
-                             (native_str("id_struct"), np.int16),
-                             (native_str("len_struct"), np.int32),
-                             (native_str("len_data"), np.int32)])
+structtag_dtypes = np.dtype([("sinc", "1S"),
+                             ("machine", "1S"),
+                             ("id_struct", np.int16),
+                             ("len_struct", np.int32),
+                             ("len_data", np.int32)])
 
 types = {"s": np.uint16, "q": np.int16, "u": np.uint16, "i": np.int16,
          "2": np.int32, "l": np.int32, "r": np.uint16, "f": np.float32,
