+++ obspy/realtime/tests/test_rttrace.py
@@ -163,7 +163,7 @@ class RtTraceTestCase(unittest.TestCase):
         Test for not using float32.
         """
         tr = read()[0]
-        tr.data = np.require(tr.data, dtype=native_str('>f4'))
+        tr.data = np.require(tr.data, dtype='>f4')
         traces = tr / 3
         rtr = RtTrace()
         for trace in traces:
