+++ obspy/db/scripts/indexer.py
@@ -139,7 +139,7 @@ def _run_indexer(options):
             p.daemon = True
             p.start()
         # connect to database
-        engine = create_engine(options.db_uri, encoding=native_str('utf-8'),
+        engine = create_engine(options.db_uri, encoding='utf-8',
                                convert_unicode=True)
         metadata = Base.metadata
         # recreate database
