+++ obspy/core/tests/test_waveform_plugins.py
@@ -540,7 +540,6 @@ class WaveformPluginsTestCase(unittest.TestCase):
             break
         else:
             self.fail('unable to get invalid file path')
-        doesnt_exist = native_str(doesnt_exist)
 
         if PY2:
             exception_type = getattr(builtins, 'IOError')

[…]
