### 2020-03-09 bc9435e
sunset py2: remove future's native_str in isinstance checks
