+++ obspy/io/json/tests/test_json.py
@@ -24,7 +24,7 @@ class JSONTestCase(unittest.TestCase):
 
     def verify_json(self, s):
         """Test an output is a string and is JSON"""
-        self.assertTrue(isinstance(s, (str, native_str)))
+        self.assertTrue(isinstance(s, str))
         j = json.loads(s)
         self.assertTrue(isinstance(j, dict))
 
