+++ obspy/io/xseed/blockette/blockette041.py
@@ -43,7 +43,7 @@ class Blockette041(Blockette):
         if isinstance(data, bytes):
             expected_length = len(data)
             data = io.BytesIO(data)
-        elif isinstance(data, (str, native_str)):
+        elif isinstance(data, str):
             raise TypeError("Data must be bytes, not string")
         # get current lookup key
         pos = data.tell()
