### 2020-03-09 ee6be4a
sunset py2: remove native_bytes usage
