### 2020-03-09 ab6e901
sunset py2: remove loads of Py2/3 if/else and compat
