+++ obspy/core/tests/test_event.py
@@ -233,10 +233,6 @@ class CatalogTestCase(unittest.TestCase):
         else:
             self.fail('unable to get invalid file path')
 
-        if PY2:
-            exception_type = getattr(builtins, 'IOError')
-        else:
-            exception_type = getattr(builtins, 'FileNotFoundError')
         exception_msg = "[Errno 2] No such file or directory: '{}'"
 
         formats = _get_entry_points(
@@ -245,7 +241,7 @@ class CatalogTestCase(unittest.TestCase):
         # plugins and also for filetype autodiscovery
         formats = [None] + list(formats)
         for format in formats:
-            with self.assertRaises(exception_type) as e:
+            with self.assertRaises(FileNotFoundError) as e:
                 read_events(doesnt_exist, format=format)
             self.assertEqual(
                 str(e.exception), exception_msg.format(doesnt_exist))
