### 2020-03-09 0cb8c04
sunset py2: remove mock compatibility layer
