+++ obspy/core/tests/test_stream.py
@@ -7,12 +7,12 @@ import platform
 import unittest
 import warnings
 from copy import deepcopy
+from unittest import mock
 
 import numpy as np
 
 from obspy import Stream, Trace, UTCDateTime, read, read_inventory
 from obspy.core.inventory import Channel, Inventory, Network, Station
-from obspy.core.compatibility import mock
 from obspy.core.stream import _is_pickle, _read_pickle, _write_pickle
 from obspy.core.util.attribdict import AttribDict
 from obspy.core.util.base import NamedTemporaryFile, _get_entry_points
