+++ obspy/core/tests/test_inventory.py
@@ -15,13 +15,13 @@ import io
 import os
 import unittest
 import warnings
+from unittest import mock
 
 import numpy as np
 from matplotlib import rcParams
 
 import obspy
 from obspy import UTCDateTime, read_inventory, read_events
-from obspy.core.compatibility import mock
 from obspy.core.util import (
     BASEMAP_VERSION, CARTOPY_VERSION, MATPLOTLIB_VERSION, PROJ4_VERSION)
 from obspy.core.util.base import _get_entry_points
