+++ obspy/core/tests/test_util_base.py
@@ -3,8 +3,8 @@ import os
 import copy
 import shutil
 import unittest
+from unittest import mock
 
-from obspy.core.compatibility import mock
 from obspy.core.util.base import (NamedTemporaryFile, get_dependency_version,
                                   download_to_file, sanitize_filename,
                                   create_empty_data_chunk, ComparingObject)
