### 2020-03-09 256ac8a
sunset py2: remove most content from core.compatibility and its usages
