+++ obspy/clients/fdsn/tests/test_client.py
@@ -25,7 +25,6 @@ import numpy as np
 import requests
 
 from obspy import UTCDateTime, read, read_inventory, Stream, Trace
-from obspy.core.compatibility import RegExTestCase
 from obspy.core.util.base import NamedTemporaryFile
 from obspy.clients.fdsn import Client, RoutingClient
 from obspy.clients.fdsn.client import build_url, parse_simple_xml
@@ -86,7 +85,7 @@ def normalize_version_number(string):
     return [l.strip() for l in repl.splitlines()]
 
 
-class ClientTestCase(RegExTestCase):
+class ClientTestCase(unittest.TestCase):
     """
     Test cases for obspy.clients.fdsn.client.Client.
     """
