+++ obspy/clients/fdsn/client.py
@@ -19,12 +19,13 @@ import textwrap
 import threading
 import warnings
 from collections import OrderedDict
+from urllib.parse import urlparse
 
 from lxml import etree
 
 import obspy
 from obspy import UTCDateTime, read_inventory
-from obspy.core.compatibility import urlparse, collections_abc
+from obspy.core.compatibility import collections_abc
 from .header import (DEFAULT_PARAMETERS, DEFAULT_USER_AGENT, FDSNWS,
                      OPTIONAL_PARAMETERS, PARAMETER_ALIASES, URL_MAPPINGS,
                      WADL_PARAMETERS_NOT_TO_BE_PARSED, DEFAULT_SERVICES,
