+++ obspy/core/tests/test_compatibility.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 import unittest
 
-from obspy.core.compatibility import py3_round
-
 
 class CompatibilityTestCase(unittest.TestCase):
-    def test_py3_round(self):
+    def test_round(self):
         """
-        Ensure py3_round rounds correctly and returns expected data types.
+        Ensure round rounds correctly and returns expected data types.
+
+        Maybe not needed after Py2 sunset? This usedd to be a test for
+        compatibility function py3_round which was only an alias for round on
+        Py3
         """
         # list of tuples; (input, ndigits, expected, excpected type)
         test_list = [
@@ -22,9 +24,9 @@ class CompatibilityTestCase(unittest.TestCase):
         for number, ndigits, expected, expected_type in test_list:
             # this is a fix for py3.4 which cannot take None as ndigits
             if ndigits is not None:
-                out = py3_round(number, ndigits)
+                out = round(number, ndigits)
             else:
-                out = py3_round(number)
+                out = round(number)
             self.assertEqual(out, expected)
             self.assertIsInstance(out, expected_type)
 
