### 2020-03-09 9a3dc91
sunset py2: switch Circle CI to Py3
