### 2020-03-09 b82474c
sunset py2: replace raise_from with Py3 builtin syntax
