### 2020-03-09 943c51b
sunset py2: remove unused imports and no-op left overs from future
