+++ obspy/core/tests/test_event.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import builtins
 import io
 import os
 import pickle
