### 2020-03-09 b32c84c
sunset py2: fix CircleCI switch to Py3
