### 2020-03-09 add4a08
sunset py2: replace with_metaclass
