### 2020-03-09 68447f6
sunset py2: remove some left over future thingies
