### 2020-03-09 f42cb41
Fix for Feature class removal from setuptools v46
The setuptools.Feature class was removed from setuptools v46.0.0.  For now,
detect that case and disable the --with-system-libs option. A better solution
would find an alternative to setuptools.Feature.
