### 2020-03-10 3f3e206
inv plotting: fix "if axes:" check for arrays
