+++ obspy/core/inventory/station.py
@@ -540,7 +540,7 @@ class Station(BaseNode):
         """
         import matplotlib.pyplot as plt
 
-        if axes:
+        if axes is not None:
             ax1, ax2 = axes
             fig = ax1.figure
         else:
