### 2020-03-10 e7edc1b
add sect_ax kwarg
