### 2020-03-11 fb55b0d
remove official support for Python 3.4 and remove it from CI
