### 2020-03-11 219b24d
Update CONTRIBUTORS.txt
see #2578
