+++ obspy/CONTRIBUTORS.txt
@@ -22,6 +22,7 @@ Crawford, Wayne
 Danecek, Peter
 van Driel, Martin
 Egdorf, Sven
+Ely, Geoffrey
 Engels, Fabian
 Ermert, Laura
 Eulenfeld, Tom
