### 2020-03-11 c40cc96
setup.py: add "python_requires" >=3.5
