### 2020-03-13 5ebcbb2
bugfix dependency check in runtests' report
