+++ obspy/scripts/runtests.py
@@ -271,7 +271,7 @@ def _create_report(ttrs, timetaken, log, server, hostname, sorted_tests,
         temp = module_.split('.')
         try:
             mod = __import__(module_,
-                             fromlist=[temp[1:]])
+                             fromlist=temp[1:])
         except ImportError:
             version_ = '---'
         else:
