### 2020-03-13 0f5b821
replace matplotlib rc text.hinting to reduce verbosity (deprecation)
