### 2020-03-13 426d4e8
add raw trace support to Kinemetrics' EVT format
ref to #2582
