+++ obspy/CONTRIBUTORS.txt
@@ -72,9 +72,9 @@ Murray-Bergquist, Louisa
 Nesbitt, Ian M.
 Nof, Ran Novitsky
 Panning, Mark P.
-Rapagnani, Giovanni
 Parker, Tom
 Pestourie, Romain
+Rapagnani, Giovanni
 Reyes, Celso
 Ringler, Adam
 Rothenhäusler, Nicolas
