### 2020-03-13 bc40666
add test case for raw=True
