### 2020-03-13 13f9b33
Fix .travis.yml
.travis.yml had an error involving something about bumpy not being installed. I though that conda installing it wasn't working so I added pip install numpy right below pip install geographiclib.
