### 2020-03-15 cd55dfd
We need to remove conda remove certifi
