+++ setup.py
@@ -787,8 +787,6 @@ def setupPackage():
             'Operating System :: OS Independent',
             'Programming Language :: Python',
             'Programming Language :: Python :: 3',
-            'Programming Language :: Python :: 3.5',
-            'Programming Language :: Python :: 3.6',
             'Programming Language :: Python :: 3.7',
             'Programming Language :: Python :: 3.8',
             'Topic :: Scientific/Engineering',
@@ -797,7 +795,7 @@ def setupPackage():
         packages=find_packages(),
         namespace_packages=[],
         zip_safe=False,
-        python_requires='>=3.5',
+        python_requires='>=3.7',
         install_requires=INSTALL_REQUIRES,
         extras_require=EXTRAS_REQUIRE,
         features=add_features(),
