+++ appveyor.yml
@@ -22,15 +22,6 @@ environment:
     # a later point release.
     # See: http://www.appveyor.com/docs/installed-software#python
 
-    - PYTHON: "C:\\Miniconda35-x64"
-      PYTHON_VERSION: "3.5"
-      PYTHON_ARCH: "64"
-      SKIP_NOTAG: "true"
-
-    - PYTHON: "C:\\Miniconda36-x64"
-      PYTHON_VERSION: "3.6"
-      PYTHON_ARCH: "64"
-
     - PYTHON: "C:\\Miniconda37-x64"
       PYTHON_VERSION: "3.7"
       PYTHON_ARCH: "64"
@@ -40,8 +31,8 @@ environment:
       PYTHON_ARCH: "64"
 
 matrix:
-    allow_failures:
-      - PYTHON_VERSION: "3.5"
+    # allow_failures:
+    #   - PYTHON_VERSION: "3.5"
     # enable fast fail strategy
     fast_finish: true
 
@@ -87,9 +78,9 @@ install:
   # pip 10.0.1 seems buggy and breaks all our appveyor builds
   - ps: |
       If ($env:PYTHON_VERSION -eq "3.8") {
-          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests jsonschema cryptography proj4
+          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy nose gdal decorator requests jsonschema cryptography proj4
       }else {
-          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy mock nose gdal decorator requests basemap jsonschema pyshp cryptography proj4
+          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy nose gdal decorator requests basemap jsonschema pyshp cryptography proj4
       }
   # somehow basemap/pyproj/proj4 can fail when PROJ_LIB is not set correctly
   # for whatever reason, see:
