### 2020-03-16 3376090
CI: pyshp, pyimgur, geographiclib should be installable via conda
