+++ .travis.yml
@@ -116,18 +116,21 @@ install:
         $BASEMAP
         pip
         pyproj
+        pyshp
         proj4
         codecov
         coverage
         certifi>=2017.4.17
         decorator
         docopt
+        geographiclib
         jsonschema
         lxml
         nose
         requests
         sqlalchemy
         cryptography
+        pyimgur
   - source activate test-environment
   # additional, optional packages that run some more tests but are not
   # available for all archs. "--no-update-dependencies" keeps other packages at
@@ -138,17 +141,6 @@ install:
   # least our builds work..
   # XXX - conda install --no-update-dependencies m2crypto || true
   # XXX - conda install --no-update-dependencies cartopy || true
-  # pyshp is safe to install, it depends on no other packages and has a noarch
-  # py6 package
-  # not a clue why conda is trying to pull in other packages when doing
-  # "conda install -c conda-forge --no-update-dependencies pyshp", because the
-  # pyshp conda package lists 0 dependencies, anyway, so use "--no-deps" to
-  # really avoid installing/updating any other packages
-  - conda install -c conda-forge --no-update-deps --no-deps pyshp || pip install pyshp
-  # install packages not available via conda
-  - pip install geographiclib
-  # current pyimgur stable release has a py3 incompatibility
-  - pip install https://github.com/megies/PyImgur/archive/py3.zip
   - pip freeze
   - pip install https://github.com/obspy/obspy_github_api/archive/0.7.0.zip
   - conda list
