+++ appveyor.yml
@@ -78,9 +78,9 @@ install:
   # pip 10.0.1 seems buggy and breaks all our appveyor builds
   - ps: |
       If ($env:PYTHON_VERSION -eq "3.8") {
-          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy nose gdal decorator requests jsonschema cryptography proj4
+          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy nose gdal decorator requests jsonschema cryptography proj4 pyimgur geographiclib
       }else {
-          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy nose gdal decorator requests basemap jsonschema pyshp cryptography proj4
+          conda install -q --yes "pip>10.0.1" numpy scipy matplotlib lxml sqlalchemy nose gdal decorator requests jsonschema cryptography proj4 pyimgur geographiclib basemap pyshp
       }
   # somehow basemap/pyproj/proj4 can fail when PROJ_LIB is not set correctly
   # for whatever reason, see:
@@ -90,9 +90,6 @@ install:
   #  This is really ugly and should ideally reverted at some later point, in
   #  case used paths / env var names change etc
   - "SET PROJ_LIB=%CONDA_PREFIX%\\Library\\share"
-  # additional dependencies
-  - "pip install pyimgur"
-  - "pip install geographiclib"
   # list package versions
   - "conda list"
 
