### 2020-03-16 #16 @lwasser
ok team. My apologies for being absent for the past few weeks. So many things have been happening. getting back to this and reading through everything again, it seems like the reviewers are generally happy with obspy now. I see that @megies did make the install a bit more clear! thank you. it looks great and i see numpy mentioned as a requirement.

So here we are left with two bigger issues --
1. How much documentation is "good enough" for pyopensci
2. Tests - how much do we ask of dev's in terms of testing frameworks when some tests fail locally. options
   * pytest markers which can work well but you are essentially skipping running some tests
   * leave it as is
   * Fix the tests (which may be quite onerous in this case to ensure they run across all user machines

 @canyon289  and @raoulcollenteur  have i represented things well above? are there other outstanding issues in this particular review?
