### 2020-03-16 360a948
Add code from PR #1718 to remove evalresp dependency
