+++ obspy/core/inventory/response.py
@@ -27,7 +27,8 @@ from obspy.core.util.obspy_types import (ComplexWithUncertainties,
                                          FloatWithUncertainties,
                                          FloatWithUncertaintiesAndUnit,
                                          ObsPyException,
-                                         ZeroSamplingRate)
+                                         ZeroSamplingRate,
+                                         Enum)
 
 from .util import Angle, Frequency
 
@@ -211,7 +212,7 @@ class PolesZerosResponseStage(ResponseStage):
                  stage_gain_frequency, input_units, output_units,
                  pz_transfer_function_type,
                  normalization_frequency, zeros, poles,
-                 normalization_factor=1.0, resource_id=None, resource_id2=None,
+                 normalization_factor=None, resource_id=None, resource_id2=None,
                  name=None, input_units_description=None,
                  output_units_description=None, description=None,
                  decimation_input_sample_rate=None, decimation_factor=None,
@@ -221,6 +222,8 @@ class PolesZerosResponseStage(ResponseStage):
         # handled by properties.
         self.pz_transfer_function_type = pz_transfer_function_type
         self.normalization_frequency = normalization_frequency
+        if normalization_factor is None:
+            normalization_factor = self.calc_normalization_factor()
         self.normalization_factor = float(normalization_factor)
         self.zeros = zeros
         self.poles = poles
@@ -319,6 +322,54 @@ class PolesZerosResponseStage(ResponseStage):
             self._pz_transfer_function_type = "DIGITAL (Z-TRANSFORM)"
         else:
             raise ValueError(msg)
+
+    def get_response(self, frequencies):
+        """
+        Produce the response curve from this stage's data for a given
+        range of frequencies
+        :param frequencies: Frequency range to get resp curve over
+        :return: The curve describing this response stage
+        """
+        # Has to be imported here for now to avoid circular imports.
+        from obspy.signal.invsim import paz_to_freq_resp
+        return paz_to_freq_resp(
+            poles=np.array(self.poles, dtype=np.complex128),
+            zeros=np.array(self.zeros, dtype=np.complex128),
+            scale_fac=self.normalization_factor,
+            frequencies=frequencies, freq=False) * self.stage_gain
+
+    def calc_normalization_factor(self):
+        """
+        Calculate the normalization factor for given poles-zeros
+
+        The norm factor A0 is calculated such that
+                           sequence_product_over_n(s - zero_n)
+                A0 * abs(------------------------------------------) === 1
+                           sequence_product_over_m(s - pole_m)
+        for s_f=i*2pi*f if the transfer function is in radians
+                i*f     if the transfer funtion is in Hertz
+        :return: the value A0
+        """
+        # code for this method provided by @Wayne_Crawford
+        if not self.normalization_frequency:
+            return None
+
+        A0 = 1.0 + (1j * 0.0)
+        # TODO: ensure that this coercion to float is valid
+        if self.transfer_function_type == "LAPLACE (HERTZ)":
+            s = 1j * float(self.normalization_frequency)
+        elif self.transfer_function_type == "LAPLACE (RADIANS/SECOND)":
+            s = 1j * 2 * pi * float(self.normalization_frequency)
+        else:
+            print("Don't know how to calculate normalization factor "
+                  "for z-transform poles and zeros!")
+            return False
+        for p in self.poles:
+            A0 *= (s - p)
+        for z in self.zeros:
+            A0 /= (s - z)
+
+        return abs(A0)
[…]
