+++ obspy/signal/invsim.py
@@ -351,7 +351,8 @@ def corn_freq_2_paz(fc, damp=0.707):
     return {'poles': poles, 'zeros': [0j, 0j], 'gain': 1, 'sensitivity': 1.0}
 
 
-def paz_to_freq_resp(poles, zeros, scale_fac, t_samp, nfft, freq=False):
+def paz_to_freq_resp(poles, zeros, scale_fac, t_samp=None, nfft=None,
+                     frequencies=None, freq=False):
     """
     Convert Poles and Zeros (PAZ) to frequency response.
 
@@ -367,18 +368,26 @@ def paz_to_freq_resp(poles, zeros, scale_fac, t_samp, nfft, freq=False):
     :param t_samp: Sampling interval in seconds
     :type nfft: int
     :param nfft: Number of FFT points of signal which needs correction
+    :type frequencies: list of float
+    :param frequencies: Discrete frequencies to get resp values for.
+    :type freq: bool
+    :param freq: If true, returns tuple of resp result with freq array input (i.e., x-values)
     :rtype: :class:`numpy.ndarray` complex128
     :return: Frequency response of PAZ of length nfft
     """
-    n = nfft // 2
     b, a = scipy.signal.ltisys.zpk2tf(zeros, poles, scale_fac)
     # a has to be a list for the scipy.signal.freqs() call later but zpk2tf()
     # strangely returns it as an integer.
     if not isinstance(a, np.ndarray) and a == 1.0:
         a = [1.0]
-    fy = 1 / (t_samp * 2.0)
-    # start at zero to get zero for offset / DC of fft
-    f = np.linspace(0, fy, n + 1)
+
+    if frequencies is None:
+        n = nfft // 2
+        fy = 1 / (t_samp * 2.0)
+        # start at zero to get zero for offset / DC of fft
+        f = np.linspace(0, fy, n + 1)
+    else:
+        f = frequencies
     _w, h = scipy.signal.freqs(b, a, f * 2 * np.pi)
     if freq:
         return h, f
