### 2020-03-16 e65097b
Fix some issues related to test cases failing
