+++ obspy/core/tests/test_response.py
@@ -329,10 +329,10 @@ class ResponseTestCase(unittest.TestCase):
         zeros = [2 + 3j, 2, 3j]
         stage = PolesZerosResponseStage(
             1, 1, 1, "", "", "LAPLACE (HERTZ)", 1, zeros, poles)
-        self.assertEqual(type(stage.zeros[0]), ComplexWithUncertainties)
-        self.assertEqual(type(stage.poles[0]), ComplexWithUncertainties)
-        self.assertEqual(stage.poles, poles)
-        self.assertEqual(stage.zeros, zeros)
+        self.assertEqual(ComplexWithUncertainties, type(stage.zeros[0]))
+        self.assertEqual(ComplexWithUncertainties, type(stage.poles[0]))
+        self.assertEqual(poles, stage.poles)
+        self.assertEqual(zeros, stage.zeros)
 
     def test_response_list_stage(self):
         """
@@ -637,8 +637,8 @@ class ResponseTestCase(unittest.TestCase):
         """
         inv = read_inventory(os.path.join(self.data_dir, "IM_I53H1_BDF.xml"))
         self.assertEqual(
-            inv[0][0][0].response.get_evalresp_response_for_frequencies([0.0]),
-            0.0 + 0.0j)
+            0.0 + 0.0j,
+            inv[0][0][0].response.get_evalresp_response_for_frequencies([0.0])[0])
         np.testing.assert_allclose(
             inv[0][0][0].response.get_evalresp_response_for_frequencies(
                 [0.1, 1.0, 10.0]),
