+++ obspy/core/inventory/response.py
@@ -129,15 +129,15 @@ class ResponseStage(ComparingObject):
         self.description = description
         self.decimation_input_sample_rate = \
             Frequency(decimation_input_sample_rate) \
-            if decimation_input_sample_rate is not None else None
+                if decimation_input_sample_rate is not None else None
         self.decimation_factor = decimation_factor
         self.decimation_offset = decimation_offset
         self.decimation_delay = \
             FloatWithUncertaintiesAndUnit(decimation_delay) \
-            if decimation_delay is not None else None
+                if decimation_delay is not None else None
         self.decimation_correction = \
             FloatWithUncertaintiesAndUnit(decimation_correction) \
-            if decimation_correction is not None else None
+                if decimation_correction is not None else None
 
     def __str__(self):
         ret = (
@@ -222,11 +222,18 @@ class PolesZerosResponseStage(ResponseStage):
         # handled by properties.
         self.pz_transfer_function_type = pz_transfer_function_type
         self.normalization_frequency = normalization_frequency
+        for i, x in enumerate(zeros):
+            if not isinstance(x, ComplexWithUncertainties):
+                zeros[i] = ComplexWithUncertainties(x)
+        self._zeros = zeros
+        for i, x in enumerate(poles):
+            if not isinstance(x, ComplexWithUncertainties):
+                poles[i] = ComplexWithUncertainties(x)
+        self._poles = poles
+        # if we don't have a normalization factor, calculate it (needs P/Z set first)
         if normalization_factor is None:
             normalization_factor = self.calc_normalization_factor()
         self.normalization_factor = float(normalization_factor)
-        self.zeros = zeros
-        self.poles = poles
         super(PolesZerosResponseStage, self).__init__(
             stage_sequence_number=stage_sequence_number,
             input_units=input_units,
@@ -255,8 +262,8 @@ class PolesZerosResponseStage(ResponseStage):
             transfer_fct_type=self.pz_transfer_function_type,
             norm_fact=self.normalization_factor,
             norm_freq=self.normalization_frequency,
-            poles=", ".join(map(str, self.poles)),
-            zeros=", ".join(map(str, self.zeros)))
+            poles=", ".join(map(str, self._poles)),
+            zeros=", ".join(map(str, self._zeros)))
         return ret
 
     def _repr_pretty_(self, p, cycle):
@@ -333,8 +340,8 @@ class PolesZerosResponseStage(ResponseStage):
         # Has to be imported here for now to avoid circular imports.
         from obspy.signal.invsim import paz_to_freq_resp
         return paz_to_freq_resp(
-            poles=np.array(self.poles, dtype=np.complex128),
-            zeros=np.array(self.zeros, dtype=np.complex128),
+            poles=np.array(self._poles, dtype=np.complex128),
+            zeros=np.array(self._zeros, dtype=np.complex128),
             scale_fac=self.normalization_factor,
             frequencies=frequencies, freq=False) * self.stage_gain
 
@@ -356,17 +363,17 @@ class PolesZerosResponseStage(ResponseStage):
 
         A0 = 1.0 + (1j * 0.0)
         # TODO: ensure that this coercion to float is valid
-        if self.transfer_function_type == "LAPLACE (HERTZ)":
+        if self.pz_transfer_function_type == "LAPLACE (HERTZ)":
             s = 1j * float(self.normalization_frequency)
-        elif self.transfer_function_type == "LAPLACE (RADIANS/SECOND)":
+        elif self.pz_transfer_function_type == "LAPLACE (RADIANS/SECOND)":
             s = 1j * 2 * pi * float(self.normalization_frequency)
         else:
             print("Don't know how to calculate normalization factor "
[…]
