### 2020-03-16 cc58fbc
Check if resp calculation valid for resp stage first
