+++ obspy/core/inventory/response.py
@@ -993,7 +993,8 @@ class Response(ComparingObject):
         stages = self.response_stages[slice(start_stage, end_stage)]
         resp = stages.pop(0).get_response(frequencies=frequencies)
         for stage in stages[1:]:
-            resp *= stage.get_response(frequencies=frequencies)
+            if 'get_response' in dir(stage):
+                resp *= stage.get_response(frequencies=frequencies)
 
         # For the scaling - run the whole chain once again with the
         # reference frequency.
@@ -1002,7 +1003,8 @@ class Response(ComparingObject):
             stages = self.response_stages[slice(start_stage, end_stage)]
             ref = stages.pop(0).get_response(frequencies=f)
             for stage in stages[1:]:
-                ref *= stage.get_response(frequencies=f)
+                if 'get_response' in dir(stage):
+                    ref *= stage.get_response(frequencies=f)
             resp *= self.instrument_sensitivity.value / np.abs(ref[0])
 
         # By now the response is in the input units of the first stage.
