### 2020-03-16 6933567
Revert to making resp curve calc fail without method definition
