### 2020-03-17 173eb15
Revert test case change, was running IDE under 2.7, causing issue
