### 2020-03-18 d532a4f
Replace get_evalresp_response with getter using freq window size
