+++ obspy/core/trace.py
@@ -2907,8 +2907,8 @@ seismometer_correction_simulation.html#using-a-resp-file>`_.
         # calculate and apply frequency response,
         # optionally prefilter in frequency domain and/or apply water level
         freq_response, freqs = \
-            response.get_evalresp_response(self.stats.delta, nfft,
-                                           output=output, **kwargs)
+            response.get_response_for_window_size(self.stats.delta, nfft,
+                                                  output=output, **kwargs)
 
         if plot:
             ax1.loglog(freqs, np.abs(data), color=color1, zorder=9)
