### 2020-03-18 ae19ec2
Remove FIR error suppression hack
