+++ obspy/core/inventory/response.py
@@ -1034,8 +1034,7 @@ class Response(ComparingObject):
         stages = self.response_stages[slice(start_stage, end_stage)]
         resp = stages.pop(0).get_response(frequencies=frequencies)
         for stage in stages[1:]:
-            if 'get_response' in dir(stage):
-                resp *= stage.get_response(frequencies=frequencies)
+            resp *= stage.get_response(frequencies=frequencies)
 
         # For the scaling - run the whole chain once again with the
         # reference frequency.
