### 2020-03-19 83b05e6
only test pickle with bytesio, remove saved pickles
