+++ obspy/core/tests/test_utcdatetime.py
@@ -1484,7 +1484,7 @@ class UTCDateTimeTestCase(unittest.TestCase):
         t = UTCDateTime(1888, 1, 2, 1, 39, 37)
         self.assertEqual(t.strftime('%Y-%m-%d'), '1888-01-02')
         t = UTCDateTime(998, 11, 9, 1, 39, 37)
-        self.assertEqual(t.strftime('%Y-%m-%d'), '0998-11-09')
+        self.assertEqual(t.strftime('%Y-%m-%d'), '998-11-09')
 
     def test_strftime_replacement(self):
         """
