### 2020-03-19 e8604d6
allow bytes/str of test_stats to run
