+++ obspy/core/tests/test_stats.py
@@ -300,10 +300,6 @@ class StatsTestCase(unittest.TestCase):
         Test the various types of strings found in the wild get converted to
         native_str type.
         """
-        # XXX check what to do here, if still needed after Py2 removal
-        raise Exception(
-            'test needs to be looked at more closely after Py2 sunset')
-        # get native bytes
         nbytes = bytes('HHZ', 'utf8')
         the_strs = ['HHZ', nbytes, u'HHZ']
 
