### 2020-03-19 2e82def
remove deprecated scipy.fix
