### 2020-03-19 bc55063
bump numpy, scipy, mpl versions
