### 2020-03-19 4c17ec1
block setup.py for older versions of python
