### 2020-03-19 f51043c
add import error for ObsPy >2.0 running on python 2
