### 2020-03-19 e6867d1
loosen strfmt test for trailing 0
