### 2020-03-19 9b518dc
Update response getters (test still fails)
