### 2020-03-19 6df8648
Fix up FIR response getter a bit more
