### 2020-03-19 91efd02
Add test case to compare by-stage the response calculations
