### 2020-03-19 c10af02
Add test case to compare by-stage the response calculations
