### 2020-03-19 9fef31c
Check MEEK in test case, fails because stage 2 is a gain stage
