### 2020-03-19 8161565
Check MEEK in test case, fails because stage 2 is a gain stage
