+++ obspy/clients/arclink/__init__.py
@@ -3,159 +3,17 @@
 obspy.clients.arclink - ArcLink/WebDC request client for ObsPy
 ==============================================================
 
-ArcLink is a distributed data request protocol usable to access archived
-waveform data in the MiniSEED or SEED format and associated meta information as
-Dataless SEED files. It has been originally founded within the German WebDC
-initiative of GEOFON_ (Geoforschungsnetz) and BGR_ (Bundesanstalt für
-Geowissenschaften und Rohstoffe). ArcLink has been designed as a "straight
-consequent continuation" of the NetDC concept originally developed by the IRIS_
-DMC. Instead of requiring waveform data via E-mail or FTP requests, ArcLink
-offers a direct TCP/IP communication approach. A prototypic web-based request
-tool is available via the WebDC homepage at http://www.webdc.eu.
+DEPRECATED -- ArcLink protocol is officially deprecated
 
-Recent development efforts within the NERIES_ (Network of Excellence of
-Research and Infrastructures for European Seismology) project focuses on
-extending the ArcLink network to all major seismic data centers within Europe
-in order to create an European Integrated Data Center (EIDAC). Currently
-(September 2009) there are four European data centers contributing to this
-network: ORFEUS_, GFZ_ (GeoForschungsZentrum), INGV_ (Istituto Nazionale di
-Geofisica e Vulcanologia), and IPGP_ (Institut de Physique du Globe de Paris).
+ArcLink protocol has been officially deprecated and some main servers have been
+shut down already. Please consider using other methods like FDSN web services
+to fetch data.  ArcLink functionality is now untested in ObsPy.
 
 :copyright:
     The ObsPy Development Team (devs@obspy.org)
 :license:
     GNU Lesser General Public License, Version 3
     (https://www.gnu.org/copyleft/lesser.html)
-
-Basic Usage
------------
-
-.. note::
-    The default client needs to open port 18002 to the host webdc.eu via TCP/IP
-    in order to download the requested data. Please make sure that no firewall
-    is blocking access to this server/port combination.
-
-.. note::
-    The ``user`` keyword in the following examples is used for identification
-    with the ArcLink server as well as for usage statistics within the data
-    center, so please provide a meaningful user id such as an email address.
-
-(1) :meth:`~obspy.clients.arclink.client.Client.get_waveforms()`: The following
-    example illustrates how to request and plot 18 seconds of all three single
-    band channels (``"EH*"``) of station Jochberg/Hochstaufen (``"RJOB"``)
-    of the Bavarian network (``"BW"``) for an seismic event around
-    2009-08-20 04:03:12 (UTC).
-
-    >>> from obspy import UTCDateTime
-    >>> from obspy.clients.arclink.client import Client
-    >>> client = Client(user='test@obspy.org')
-    >>> t = UTCDateTime("2009-08-20 04:03:12")
-    >>> st = client.get_waveforms("BW", "RJOB", "", "EH*", t - 3, t + 15)
-    >>> st.plot()  # doctest: +SKIP
-
-    Waveform data fetched from an ArcLink node is converted into an ObsPy
-    :class:`~obspy.core.stream.Stream` object. The seismogram is truncated by
-    the ObsPy client to the actual requested time span, as ArcLink internally
[…]
