+++ obspy/clients/arclink/tests/test_client.py
@@ -1,783 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-The obspy.clients.arclink.client test suite.
-"""
-import io
-import operator
-import unittest
-import warnings
-
-import numpy as np
-
-from obspy import read
-from obspy.clients.arclink import Client
-from obspy.clients.arclink.client import ArcLinkException
-from obspy.core.utcdatetime import UTCDateTime
-from obspy.core.util import AttribDict, NamedTemporaryFile
-
-
-class ClientTestCase(unittest.TestCase):
-    """
-    Test cases for obspy.clients.arclink.client.Client.
-    """
-    def test_get_waveform(self):
-        """
-        Tests get_waveforms method.
-        """
-        # example 1
-        client = Client(user='test@obspy.org')
-        start = UTCDateTime(2010, 1, 1)
-        end = start + 1
-        stream = client.get_waveforms('BW', 'MANZ', '', 'EH*', start, end)
-        self.assertEqual(len(stream), 3)
-        for trace in stream:
-            self.assertLessEqual(trace.stats.starttime, start)
-            self.assertGreaterEqual(trace.stats.endtime, end)
-            self.assertEqual(trace.stats.network, 'BW')
-            self.assertEqual(trace.stats.station, 'MANZ')
-            self.assertEqual(trace.stats.location, '')
-            self.assertEqual(trace.stats.channel[0:2], 'EH')
-        # example 2
-        client = Client(user='test@obspy.org')
-        start = UTCDateTime("2010-12-31T23:59:50.495000Z")
-        end = start + 100
-        stream = client.get_waveforms('GE', 'APE', '', 'BHE', start, end)
-        # seems the response of the server has changed, it contains two traces
-        # with an overlap now but the data is consistent and a cleanup merge
-        # works
-        stream.merge(-1)
-        self.assertEqual(len(stream), 1)
-        trace = stream[0]
-        self.assertLessEqual(trace.stats.starttime, start)
-        self.assertGreaterEqual(trace.stats.endtime, end)
-        self.assertEqual(trace.stats.network, 'GE')
-        self.assertEqual(trace.stats.station, 'APE')
-        self.assertEqual(trace.stats.location, '')
-        self.assertEqual(trace.stats.channel, 'BHE')
-
-    def test_get_waveform_no_routing(self):
-        """
-        Tests routing parameter of get_waveforms method.
-        """
-        # 1 - requesting BW data w/o routing on webdc.eu
-        client = Client(user='test@obspy.org')
-        start = UTCDateTime(2008, 1, 1)
-        end = start + 1
-        self.assertRaises(ArcLinkException, client.get_waveforms, 'BW', 'MANZ',
-                          '', 'EH*', start, end, route=False)
-        # 2 - requesting BW data w/o routing directly from BW ArcLink node
-        client = Client(host='erde.geophysik.uni-muenchen.de', port=18001,
-                        user='test@obspy.org')
-        start = UTCDateTime(2008, 1, 1)
-        end = start + 1
-        stream = client.get_waveforms('BW', 'MANZ', '', 'EH*', start, end,
-                                      route=False)
-        for trace in stream:
-            self.assertLessEqual(trace.stats.starttime, start)
-            self.assertGreaterEqual(trace.stats.endtime, end)
-            self.assertEqual(trace.stats.network, 'BW')
-            self.assertEqual(trace.stats.station, 'MANZ')
-            self.assertEqual(trace.stats.location, '')
-            self.assertEqual(trace.stats.channel[0:2], 'EH')
-
-    def test_delayed_request(self):
-        """
-        """
-        # initialize client with 0.1 delay
[…]
