+++ obspy/clients/arclink/tests/__init__.py
@@ -1,18 +1,4 @@
 # -*- coding: utf-8 -*-
-import unittest
-
-from obspy.core.util import add_doctests, add_unittests
-
-
-MODULE_NAME = "obspy.clients.arclink"
-
-
-def suite():
-    suite = unittest.TestSuite()
-    add_doctests(suite, MODULE_NAME)
-    add_unittests(suite, MODULE_NAME)
-    return suite
-
-
-if __name__ == '__main__':
-    unittest.main(defaultTest='suite')
+"""
+ArcLink has been deprecated and all tests have been removed.
+"""

[…]
