### 2020-03-20 4fff8f2
Implement basic methods for FIR and IIR response getting
