+++ obspy/signal/invsim.py
@@ -380,7 +380,13 @@ def paz_to_freq_resp(poles, zeros, scale_fac, t_samp=None, nfft=None,
     # strangely returns it as an integer.
     if not isinstance(a, np.ndarray) and a == 1.0:
         a = [1.0]
+    # this turns the paz values into an IIR specification, so we will
+    # use the IIR conversion method to produce the response
+    return iir_to_freq_response(b, a, t_samp, nfft, frequencies, freq)
 
+
+def iir_to_freq_response(numer, denom, t_samp=None, nfft=None,
+                         frequencies=None, freq=False):
     if frequencies is None:
         n = nfft // 2
         fy = 1 / (t_samp * 2.0)
@@ -388,7 +394,22 @@ def paz_to_freq_resp(poles, zeros, scale_fac, t_samp=None, nfft=None,
         f = np.linspace(0, fy, n + 1)
     else:
         f = frequencies
-    _w, h = scipy.signal.freqs(b, a, f * 2 * np.pi)
+    _w, h = scipy.signal.freqs(numer, denom, f * 2 * np.pi)
+    if freq:
+        return h, f
+    return h
+
+
+def fir_to_freq_response(numer, denom, t_samp=None, nfft=None,
+                         frequencies=None, freq=False):
+    if frequencies is None:
+        n = nfft // 2
+        fy = 1 / (t_samp * 2.0)
+        # start at zero to get zero for offset / DC of fft
+        f = np.linspace(0, fy, n + 1)
+    else:
+        f = frequencies
+    _w, h = scipy.signal.freqz(numer, denom, f * 2 * np.pi)
     if freq:
         return h, f
     return h
