### 2020-03-20 2d4acf8
add linux specific strftime
