### 2020-03-21 f44af91
Committing current state for review, mostly minor changes
