+++ obspy/core/inventory/response.py
@@ -177,6 +177,11 @@ class ResponseStage(ComparingObject):
     def _repr_pretty_(self, p, cycle):
         p.text(str(self))
 
+    def get_response(self, frequencies):
+        # if a response stage isn't a subclass then it's likely a gain stage.
+        # in that case just return an array of the constant gain
+        return np.ones_like(frequencies) * self.stage_gain
+
 
 class PolesZerosResponseStage(ResponseStage):
     """
@@ -565,13 +570,14 @@ class CoefficientsTypeResponseStage(ResponseStage):
         # sure that the phase returned is equivalent to evalpres.
         amp = np.abs(resp)
         phase = np.radians(np.unwrap(np.angle(resp, deg=False))) / np.pi
+        if self.cf_transfer_function_type == "DIGITAL":
+            phase *= 0.
 
         # Normalize the amplitude with the given sensitivity value and
         # frequency. I'm not sure this is entirely correct, as the digitizer
         # will likely just apply the FIR filter and send the data along. But
         # evalresp does this and thus so do we.
         amp *= self.stage_gain / gain_freq_amp
-
         final_resp = np.empty_like(resp)
         final_resp.real = amp * np.cos(phase)
         final_resp.imag = amp * np.sin(phase)
@@ -743,6 +749,7 @@ class FIRResponseStage(ResponseStage):
         self._coefficients = new_values
 
     def get_response(self, frequencies):
+        from obspy.signal.invsim import fir_to_freq_resp
         sr = self.decimation_input_sample_rate
         frequencies = frequencies / sr * np.pi * 2.0
         if self.symmetry == 'ODD':
@@ -752,28 +759,9 @@ class FIRResponseStage(ResponseStage):
         else:
             # This is the full case
             coefficients = self.coefficients
-        resp = scipy.signal.freqz(
-            b=coefficients, a=[1.], worN=frequencies)[1]
-        gain_freq_amp = np.abs(scipy.signal.freqz(
-            b=coefficients, a=[1.],
-            worN=[self.stage_gain_frequency])[1])
-        resp = resp.conjugate()
-        # evalresp is a bit funny in how it defines the phase. Here we make
-        # sure that the phase returned is equivalent to evalpres.
-        amp = np.abs(resp)
-        phase = np.radians(np.unwrap(np.angle(resp, deg=False))) / np.pi ** 7
 
-        # Normalize the amplitude with the given sensitivity value and
-        # frequency. I'm not sure this is entirely correct, as the digitizer
-        # will likely just apply the FIR filter and send the data along. But
-        # evalresp does this and thus so do we.
-        # amp *= self.stage_gain / gain_freq_amp
-        amp *= self.stage_gain
-        final_resp = np.empty_like(resp)
-        final_resp.real = amp
-        final_resp.imag = amp * 0.
-
-        return final_resp
+        return fir_to_freq_resp(coefficients, [1.],
+                                frequencies=frequencies) * self.stage_gain
 
 
 class PolynomialResponseStage(ResponseStage):
