+++ .github/workflows/test_condarc.yml
@@ -1,0 +1,4 @@
+channels:
+  - conda-forge
+  - defaults
+auto_activate_base: false
