+++ .github/workflows/runtests.yml
@@ -1,0 +1,65 @@
+name: validate
+on: [push]
+
+jobs:
+  # Simply applies flake8 to the codebase
+  lint_code:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v1
+
+      - name: Setup conda
+        uses: s-weigand/setup-conda@v1
+        with:
+          python-version: 3.7
+
+      - name: install linting packages
+        run: pip install flake8
+
+      - name: flake8
+        run: flake8 .
+
+  # Runs the tests on combinations of the supported python/os matrix.
+  test_code:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: [3.6, 3.7]
+
+    steps:
+      - uses: actions/checkout@v1
+
+      - name: Setup conda
+        uses: goanpeca/setup-miniconda@v1
+        with:
+          miniconda-version: 'latest'
+          python-version: ${{ matrix.python-version }}
+          activate-environment: test
+          environment-file: .github/test_conda_env.yml
+          condarc-file: .github/test_condarc.yml
+
+      - name: install obsplus
+        shell: bash -l {0}
+        run: |
+          pip install -e .
+
+      - name: print package info
+        shell: bash -l {0}
+        run: |
+          conda info -a
+          conda list
+
+      # Runs test suite and calculates coverage
+      - name: run test suite
+        shell: bash -l {0}
+        run: |
+          obspy-runtests
+
+
+
+# This is a very useful step for debugging, it allows you to ssh into the CI
+# machine (https://github.com/marketplace/actions/debugging-with-tmate)
+#- name: Setup tmate session
+#  uses: mxschmitt/action-tmate@v1
