### 2020-03-23 03e30ee
Switched to one digital filter case.
