### 2020-03-23 02bed44
Switched to one digital filter case.
