### 2020-03-23 e475682
More attempts to fix FIR and coeff resp issues
