+++ obspy/core/tests/test_response.py
@@ -116,7 +116,7 @@ class ResponseTestCase(unittest.TestCase):
 
     def test_get_response_regression(self):
         units = ["DISP", "VEL", "ACC"]
-        filenames = ["IRIS_single_channel_with_response", "XM.05", "IU_ANMO_00_BHZ", "AU.MEEK"]
+        filenames = ["AU.MEEK", "IRIS_single_channel_with_response", "XM.05", "IU_ANMO_00_BHZ"]
 
         for filename in filenames:
             xml_filename = os.path.join(self.data_dir,
@@ -128,25 +128,24 @@ class ResponseTestCase(unittest.TestCase):
 
             for unit in units:
                 # Full response.
-                xml_resp = resp.get_evalresp_response_for_frequencies(
+                evrs_resp = resp.get_evalresp_response_for_frequencies(
                     frequencies=freqs, output=unit)
                 new_resp = resp.get_response(
                     frequencies=freqs, output=unit)
 
-                np.testing.assert_allclose(np.abs(xml_resp),
+                np.testing.assert_allclose(np.abs(evrs_resp),
                                            np.abs(new_resp), rtol=1E-5)
                 # Phase starts to differ slightly before Nyquist and quite a
                 # bit after. Evalresp appears to have some Gibb's artifacts
                 # and scipy's solution does look better.
                 np.testing.assert_allclose(
-                    np.unwrap(np.angle(xml_resp))[:800],
+                    np.unwrap(np.angle(evrs_resp))[:800],
                     np.unwrap(np.angle(new_resp))[:800],
                     rtol=1E-2, atol=2E-2)
             print("passed:", xml_filename)
 
     def test_get_response_per_stage(self):
-        from obspy.core.inventory import FIRResponseStage
-        filenames = ["IRIS_single_channel_with_response", "XM.05", "IU_ANMO_00_BHZ", "AU.MEEK"]
+        filenames = ["AU.MEEK", "IRIS_single_channel_with_response", "XM.05", "IU_ANMO_00_BHZ"]
         units = ["DISP", "VEL", "ACC"]
 
         for filename in filenames:
@@ -158,13 +157,11 @@ class ResponseTestCase(unittest.TestCase):
             print(xml_filename)
             for unit in units:
                 for x in range(1, len(resp.response_stages)+1):
-                    print(type(resp.response_stages[x - 1]))
                     """
                     if not isinstance(resp.response_stages[x-1], FIRResponseStage):
-                        print("skipping...")
                         continue
                     """
-
+                    print("Type of stage ", str(x) + ":", type(resp.response_stages[x-1]))
                     xml_resp = resp.get_evalresp_response_for_frequencies(
                         frequencies=freqs, start_stage=x, end_stage=x, output=unit)
                     new_resp = resp.get_response(
