### 2020-03-25 9bfa573
strftime: introspective approach to get rid of %Y problems < year 1000
