### 2020-03-25 2000bc8
Update the digital coefficients stage based on code from Ringler
