### 2020-03-25 72ec1ce
Add test for PZ normalization calculation
