### 2020-03-25 ddd9202
Very minor tweaks to coeff response code
