### 2020-03-26 f05c176
I implemented the FIR filter and cleaned up some flake8 issues.
