+++ obspy/core/inventory/response.py
@@ -32,6 +32,7 @@ from obspy.core.util.obspy_types import (ComplexWithUncertainties,
 
 from .util import Angle, Frequency
 
+
 class ResponseStage(ComparingObject):
     """
     From the StationXML Definition:
@@ -128,15 +129,15 @@ class ResponseStage(ComparingObject):
         self.description = description
         self.decimation_input_sample_rate = \
             Frequency(decimation_input_sample_rate) \
-                if decimation_input_sample_rate is not None else None
+            if decimation_input_sample_rate is not None else None
         self.decimation_factor = decimation_factor
         self.decimation_offset = decimation_offset
         self.decimation_delay = \
             FloatWithUncertaintiesAndUnit(decimation_delay) \
-                if decimation_delay is not None else None
+            if decimation_delay is not None else None
         self.decimation_correction = \
             FloatWithUncertaintiesAndUnit(decimation_correction) \
-                if decimation_correction is not None else None
+            if decimation_correction is not None else None
 
     def __str__(self):
         ret = (
@@ -215,8 +216,8 @@ class PolesZerosResponseStage(ResponseStage):
                  stage_gain_frequency, input_units, output_units,
                  pz_transfer_function_type,
                  normalization_frequency, zeros, poles,
-                 normalization_factor=None, resource_id=None, resource_id2=None,
-                 name=None, input_units_description=None,
+                 normalization_factor=None, resource_id=None,
+                 resource_id2=None, name=None, input_units_description=None,
                  output_units_description=None, description=None,
                  decimation_input_sample_rate=None, decimation_factor=None,
                  decimation_offset=None, decimation_delay=None,
@@ -233,7 +234,8 @@ class PolesZerosResponseStage(ResponseStage):
             if not isinstance(x, ComplexWithUncertainties):
                 poles[i] = ComplexWithUncertainties(x)
         self._poles = poles
-        # if we don't have a normalization factor, calculate it (needs P/Z set first)
+        # if we don't have a normalization factor
+        # calculate it (needs P/Z set first)
         if normalization_factor is None:
             normalization_factor = self.calc_normalization_factor()
         self.normalization_factor = float(normalization_factor)
@@ -520,8 +522,8 @@ class CoefficientsTypeResponseStage(ResponseStage):
 
     def get_response(self, frequencies):
         """
-        Produce the response curve from this coefficient response stage for a range
-        of frequencies
+        Produce the response curve from this coefficient
+        response stage for a range of frequencies
         :param frequencies: Frequency range to get resp curve over
         :return: The curve describing this response stage
         """
@@ -537,16 +539,19 @@ class CoefficientsTypeResponseStage(ResponseStage):
         # digital, this may not be the case
         if self.cf_transfer_function_type == "DIGITAL":
             if len(self.denominator) == 0:
-                resp = scipy.signal.freqz(b=self.numerator, a=[1.], worN=frequencies)[1]
+                resp = scipy.signal.freqz(b=self.numerator,
+                                          a=[1.], worN=frequencies)[1]
 
                 gain_freq_amp = np.abs(scipy.signal.freqz(
-                    b=self.numerator, a=[1.], worN=[self.stage_gain_frequency])[1])
+                    b=self.numerator, a=[1.],
+                    worN=[self.stage_gain_frequency])[1])
             else:
-                # evalresp handles this a bit oddly, and we can't just use freqs
-                # or freqz to calculate our terms here. we get the numerator and
-                # denominator and do the math on them in their representation
-                # as magnitude and phase rather than standard complex format
-                w = frequencies  # rename to be concise and match the math conventions
+                # evalresp handles this a bit oddly, and we can't just
[…]
