### 2020-03-31 fb1d985
31.03.2020: fixed and improved issues raised by megies
