+++ misc/docs/source/tutorial/code_snippets/trigger_tutorial_aic.py
@@ -6,4 +6,4 @@ trace_s = trace.slice(UTCDateTime('1970-01-01T01:00:31.6'),
                       UTCDateTime('1970-01-01T01:00:34.3'))
 p_idx, aic_f = aic(trace_s.data)
 print(p_idx / df)
-print(UTCDateTime('1970-01-01T01:00:31.6')+(p_idx / df))
+print(UTCDateTime('1970-01-01T01:00:31.6') + (p_idx / df))
