+++ CHANGELOG.txt
@@ -1,5 +1,46 @@
-master:
+1.2.1 (doi: 10.5281/zenodo.3706479)
+===================================
+
+Changes:
+ - fix an installation issue with pip and setuptools version 46 (see #2578)
+ - fix response plots when providing `axes=...` with a numpy array of Axes
+   instances (see #2579)
+
+1.2.0 (doi: 10.5281/zenodo.3674646)
+===================================
+
+Work on this release was in parts and among others supported by the following
+institutions/companies and grants (in alphabetical order):
+ - Earthquake Commision of New Zealand (EQC), grant 18/753
+ - École et Observatoire des Sciences de la Terre - Université de Strasbourg
+ - ETH Zürich
+ - Friedrich-Schiller-Universität Jena
+ - Geoscience Australia
+ - Incorporated Research Institutions for Seismology (IRIS), NSF (SAGE) award
+   :: EAR-1851048
+ - Institut de Physique du Globe de Strasbourg
+ - Institut de Physique du Globe de Paris
+ - Institutions for Seismology (IRIS) through the PASSCAL Instrument Center at
+   New Mexico Tech. The facilities of the IRIS Consortium are supported by the
+   National Science Foundation under Cooperative Agreement EAR-1261681 and the
+   DOE National Nuclear Security Administration.
+ - Istituto Nazionale di Geofisica e Vulcanologia, Osservatorio Etneo (Italy),
+   Allegato B2 DPC-INGV 2012-2021 Task 10
+ - Ludwig-Maximilians-Universität München
+ - National Institute for Occupational Safety and Health
+ - Royal Netherlands Meteorological Institute (KNMI)
+ - School of Geography, Environment and Earth Sciences, Victoria University of
+   Wellington
+ - The European Union’s Horizon 2020 research and innovation programme under
+   the ChEESE project, grant agreement No. 823844
+ - The Royal Observatory of Belgium
+ - U.S. Geological Survey
+
+Changes:
  - obspy.core:
+   * Inventory objects have been adapted to StationXML 1.1 for details on
+     changes see #2510 and
+     https://github.com/FDSN/StationXML/blob/master/Changes.md
    * Fixed import of custom plugins (see #2423)
    * Fixed "==" comparison for Stream and Trace which was very slow in case of
      traces with >1e6 samples (see #2377)
@@ -14,12 +55,12 @@ master:
    * UTCDateTime objects can now accept hour, minute, second, and microsecond
      values greater than their normal limits by setting the strict keyword
      argument to False (see #2232).
-   * fixed UTCDateTime(..., julday=366) for non-leap years. This was returning
+   * Fixed UTCDateTime(..., julday=366) for non-leap years. This was returning
      January 1st of the next year in case of non-leap years being used. Now it
      properly raises an out-of-bounds ValueError (see #2369)
    * When reading StationXML/SC3ML, make sure to properly read empty string
      fields as empty strings instead of "None" (see #2519 and #2527)
-   * better ISO8601 detection for UTCDateTime objects and UTCDateTime(...,
+   * Better ISO8601 detection for UTCDateTime objects and UTCDateTime(...,
      iso8601=False) now completely disables ISO8601 handling (see #2447)
    * Added replace method to UTCDateTime class (see #2077).
    * Added remove method to Inventory class (see #2088).
@@ -58,6 +99,8 @@ master:
    * Fix a bug that was causing an exception being raised in `Response.plot()`
      because a float was being passed down to numpy.linspace as length of array
      (see #2533)
+   * Geographic select of inventory/network/station (see #2515)
+   * Select traces in Stream based on an Inventory (see #2531)
  - obspy.clients.fdsn:
    * Add new `_discover_services` boolean flag to the Client, which allows the
[…]
