### 2020-03-31 f43c7dd
31.03.2020: reverted CHANGELOG.txt to original branch file. added the simple AIC mod under latest version in CHANGELOG
