### 2020-03-31 df98506
trying to solve assertion error on server (local works perfect)
