### 2020-04-01 8e44021
removed numpy cf in txt format. using numpy.assert_allclose for CF array comparison
