### 2020-04-03 91432d1
Added __eq__ to ellipse() class
