### 2020-04-03 90f3f3d
FLAKE8 of ellipse.py and completed one error message
