### 2020-04-03 72b7f45
Added test for bad covariance values in s_file
