### 2020-04-05 #16 @canyon289
@lwasser Apologies on my late reply

That list looks good. The one other thing I would add is whether the master branch of CI should be failing or not. Basically should there be a green badge for CI on the README or not.
