### 2020-04-06 f22779e
lightened tests over CF
