+++ .github/workflows/coverage.yml
@@ -20,23 +20,15 @@ jobs:
           environment-file: .github/test_conda_env.yml
           condarc-file: .github/test_condarc.yml
 
-      - name: install obsplus
+      - name: install obspy
         shell: bash -l {0}
         run: |
-          pip install -e .[dev]
-
-      - name: download test data
-        shell: bash -l {0}
-        run: |
-          if [ ! -d ~/opsdata ]; then
-            git clone https://github.com/d-chambers/opsdata_test
-            mv opsdata_test ~/opsdata
-          fi
+          pip install -e .
 
       - name: run test suite
         shell: bash -l {0}
         run: |
-          pytest -s --cov obsplus --cov-report=xml
+          obspy-runtests
 
       - name: upload coverage
         uses: codecov/codecov-action@v1
