### 2020-04-06 a5814e8
remove other CI services
