+++ .circleci/config.yml
@@ -1,24 +1,0 @@
-# CircleCI is used exclusively to test code formatting.
-# Thus ObsPy does not even have to be installed.
-version: 2
-jobs:
-  build:
-    docker:
-      - image: continuumio/miniconda3:latest
-
-    working_directory: ~/repo
-
-    steps:
-      - checkout
-
-      - run:
-          name: install dependencies
-          command: |
-            /opt/conda/bin/conda install -y flake8 pep8-naming
-
-      - run:
-          name: run linting
-          command: |
-            echo 'verbosity = 1' >> .flake8
-            /opt/conda/bin/flake8 --version
-            /opt/conda/bin/flake8 obspy
