### 2020-04-07 9eafb6a
Fix azimuth read type
