+++ .github/workflows/runtests.yml
@@ -4,8 +4,8 @@ on: pull_request
 # Set globally scoped environmental variables.
 env:
   ISSUE_NUMBER: ${{ github.event.pull_request.number }}
-  PR_URL: github.com/obspy/obspy/pull/${{ github.event.pull_request.number }}
-  CI_URL: github.com/obspy/obspy/actions/runs/${{ github.run_id }}
+  PR_URL: https://github.com/obspy/obspy/pull/${{ github.event.pull_request.number }}
+  CI_URL: https://github.com/obspy/obspy/actions/runs/${{ github.run_id }}
   GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
 
