### 2020-04-12 577617c
use obspy_github_api cli
