### 2020-04-15 4fd28e3
remove deprecated domain argument in correlate, remove corresponding test
