### 2020-04-15 b265321
add failing deepcopy test
