### 2020-04-15 bcc1fbe
add failing test for pickle
