### 2020-04-15 2a5e01b
reuse code from attribdict
