### 2020-04-15 6c53b10
fix stats deepcopy
* remove __deepcopy__ from AttribDict
* now Stats will use __get_state__, __set_state__ mechanism for deepcopy
