### 2020-04-15 9adca5e
fix deepcopy test of AttribDict
