+++ obspy/core/tests/test_util_attribdict.py
@@ -249,14 +249,14 @@ class AttribDictTestCase(unittest.TestCase):
 
     def test_deepcopy(self):
         """
-        Tests __deepcopy__ method of AttribDict.
+        Tests deepcopy of AttribDict.
         """
         class MyAttribDict(AttribDict):
             defaults = {'test': 1}
 
         ad = MyAttribDict()
         ad.muh = 2
-        ad2 = ad.__deepcopy__()
+        ad2 = ad.copy()
         self.assertEqual(ad2.test, 1)
         self.assertEqual(ad2.muh, 2)
 
