### 2020-04-15 8c0ae13
add test pickling AttribDict with defaults
