### 2020-04-15 74b4d0b
remove __getstate__, __setstate__ methods in AttribDict
