### 2020-04-15 43b6198
update RASPISHAKE URL mapping to use https
