+++ CHANGELOG.txt
@@ -1,4 +1,4 @@
-1.2.2 (doi: )
+master
 ===================================
 
 Changes:
