+++ CHANGELOG.txt
@@ -1,5 +1,5 @@
 master
-===================================
+======
 
 Changes:
  - obspy.clients.fdsn:
