### 2020-04-20 468a05d
update RASPISHAKE URL mapping to use https (#2603)
* update RASPISHAKE URL mapping to use https

Co-authored-by: Tobias Megies <megies@users.noreply.github.com>
