### 2020-04-22 3b09841
fix mpl and install obspy in env file
