+++ .github/workflows/runtests.yml
@@ -92,7 +92,6 @@ jobs:
       - name: upload coverage
         uses: codecov/codecov-action@v1
         with:
-          file: .coverage
           flags: unittests
           name: codecov-umbrella
           fail_ci_if_error: false
