### 2020-04-29 e4aeb46
docs: mention and link Chinese translated tutorial pdf
