### 2020-05-05 3e0c645
fixed bug of not handling HTTPException
