### 2020-05-05 068ee7a
added test for handling HTTPException
