+++ obspy/CONTRIBUTORS.txt
@@ -96,6 +96,7 @@ Trani, Luca
 Uieda, Leonardo
 Walker, Andrew
 Walther, Marcus
+Wang, Yinzhi
 Wassermann, Joachim
 van Wijk, Kasper
 Williams, Mark C.
