### 2020-05-06 fd4d60c
EIDA routing: fix an issue with a mass bulk request
when requesting a bad time window for an otherwise valid station
