+++ obspy/clients/fdsn/tests/test_mass_downloader.py
@@ -20,9 +20,10 @@ import logging
 import os
 import shutil
 from socket import timeout as socket_timeout
+import sys
 import tempfile
 import unittest
-import sys
+
 if sys.version_info.major == 2:
     from httplib import HTTPException
 else:
