+++ CHANGELOG.txt
@@ -10,6 +10,7 @@ Changes:
      data when requesting an invalid, out-of-epochs time window for a valid
      station (see #2611)
    * update RASPISHAKE URL mapping to use https
+   * fix a bug of not handling HTTPException in mass_downloader (see #2606)
  - obspy.clients.neic:
    * Make client socket blocking (see #2617)
  - obspy.io.nordic:
