### 2020-05-11 b562076
Modified codes according to moderator suggestions
Removed nested "try"s in nordic.core._write_hyp_error_line()
Added rtol attribute to nordic.ellipse.Ellipse (tolerance in __eq__)
