### 2020-05-20 e78c70e
Remove neic client socket setblocking(0) call
