### 2020-05-22 ba9b817
Add note in changelog for next release
