+++ CHANGELOG.txt
@@ -1,3 +1,6 @@
+ - obspy.clients.neic:
+   * Remove neic client socket setblocking(0) call
+
 1.2.1 (doi: 10.5281/zenodo.3706479)
 ===================================
 
