### 2020-05-22 4e8d858
Update referenced maintenance branch in CONTRIBUTING.md
