### 2020-05-26 3f8253a
signal: first try at PAZ fitting from FAP info
