### 2020-05-27 4f056f6
avoid introducing a new namespace alias, especially since its for only one function
