### 2020-06-01 0f24dd1
I added bounds to the inversion and the option to give initial guesses on poles and zeros.
