### 2020-06-03 41bb2ae
Adds pyptoject.toml to fulfill PEP518
Signed-off-by: Damian Kula <dkula@unistra.fr>
