### 2020-06-04 4102cf7
Relaxes numpy version, removes other build dependencies than np and setuptools
Signed-off-by: Damian Kula <dkula@unistra.fr>
