### 2020-06-04 f889c41
minor cosmetics in pyproject.toml
