### 2020-06-09 d8396f5
Method get_waveforms_bulk added
