### 2020-06-10 b3fe401
backport, replace ccrs.Geodetic with PlateCarree
