### 2020-06-10 ad55b20
slclient: move logging.basciConfig() call out of library code
