+++ obspy/clients/seedlink/slclient.py
@@ -110,7 +110,6 @@ class SLClient(object):
         numeric_level = getattr(logging, loglevel.upper(), None)
         if not isinstance(numeric_level, int):
             raise ValueError('Invalid log level: %s' % loglevel)
-        logging.basicConfig(level=numeric_level)
         logger.setLevel(numeric_level)
 
         self.verbose = 0
@@ -353,5 +352,10 @@ class SLClient(object):
             traceback.print_exc()
 
 
-if __name__ == '__main__':
+def main():
+    logging.basicConfig()
     SLClient.main(sys.argv)
+
+
+if __name__ == '__main__':
+    main()
