### 2020-06-10 56ba838
Adds build backend to pyptoject.toml
Signed-off-by: Damian Kula <dkula@unistra.fr>
