### 2020-06-10 69d7b59
Attempted Unit Test for get_waveforms_bulk
