### 2020-06-10 b20272a
Fixed attributes in bulk test
