### 2020-06-11 0ba288c
move call to logging.basciConfig() back to its original place
