### 2020-06-16 983646b
rudimentary unit test for issue #2616
