### 2020-06-16 2ff9876
More concise unit test
