### 2020-06-18 be723de
Update sds.py
Adding `TypeError` in `get_all_stations` allows to have weird channels name in filenames like ".../1967/XX/WEIRD/D-6.D" without crashing ...
