### 2020-06-18 b63039e
be a little bit more specific in catching an exception
