+++ obspy/clients/filesystem/tests/test_sds.py
@@ -238,25 +238,25 @@ class SDSTestCase(unittest.TestCase):
                 ["CD", "ZZZ3", "00", "BHN", t + 80,t + 100],
                 ["CD", "ZZZ3", "00", "BHE", t + 120,t + 140]
             ]
-            st = Client(temp_sds.tempdir)
-            ret_val = st.get_waveforms_bulk(chunks)
+            client = Client(temp_sds.tempdir)
+            st = client.get_waveforms_bulk(chunks)
 
             for _i in range(6):
                 if _i <= 2 :
-                    self.assertEqual(ret_val[_i].stats.network, "AB")
-                    self.assertEqual(ret_val[_i].stats.station, "XYZ")
-                    self.assertEqual(ret_val[_i].stats.location, "")
+                    self.assertEqual(st[_i].stats.network, "AB")
+                    self.assertEqual(st[_i].stats.station, "XYZ")
+                    self.assertEqual(st[_i].stats.location, "")
                 elif _i >= 2 :
-                    self.assertEqual(ret_val[_i].stats.network, "CD")
-                    self.assertEqual(ret_val[_i].stats.station, "ZZZ3")
-                    self.assertEqual(ret_val[_i].stats.location, "00")
+                    self.assertEqual(st[_i].stats.network, "CD")
+                    self.assertEqual(st[_i].stats.station, "ZZZ3")
+                    self.assertEqual(st[_i].stats.location, "00")
 
-            self.assertEqual(ret_val[0].stats.channel, "HHZ")
-            self.assertEqual(ret_val[1].stats.channel, "HHN")
-            self.assertEqual(ret_val[2].stats.channel, "HHE")
-            self.assertEqual(ret_val[3].stats.channel, "BHZ")
-            self.assertEqual(ret_val[4].stats.channel, "BHN")
-            self.assertEqual(ret_val[5].stats.channel, "BHE")
+            self.assertEqual(st[0].stats.channel, "HHZ")
+            self.assertEqual(st[1].stats.channel, "HHN")
+            self.assertEqual(st[2].stats.channel, "HHE")
+            self.assertEqual(st[3].stats.channel, "BHZ")
+            self.assertEqual(st[4].stats.channel, "BHN")
+            self.assertEqual(st[5].stats.channel, "BHE")
 
 
     def test_get_all_stations_and_nslc(self):
