### 2020-06-21 264a83b
fixed PEP8 breaches
