### 2020-06-23 0950fa5
Imports DistutilsSetupError directly from distutils instead from numpy
Signed-off-by: Damian Kula <dkula@unistra.fr>
