+++ obspy/core/event/catalog.py
@@ -632,13 +632,13 @@ class Catalog(object):
             cat = read_events()
             cat.plot(projection="local")
 
-        Combining a station and event plot (uses basemap):
+        Combining a station and event plot:
 
         >>> from obspy import read_inventory, read_events
         >>> inv = read_inventory()
         >>> cat = read_events()
-        >>> fig = inv.plot(method=basemap, show=False)  # doctest:+SKIP
-        >>> cat.plot(method=basemap, fig=fig)  # doctest:+SKIP
+        >>> fig = inv.plot(show=False)  # doctest:+SKIP
+        >>> cat.plot(fig=fig)  # doctest:+SKIP
 
         .. plot::
 
