### 2020-06-24 6221481
starting to cleanup/upgrade doc building process using sphinx
