+++ misc/docs/source/conf.py
@@ -12,9 +12,10 @@ import obspy
 import os
 import sys
 
-import cartopy
-
 import matplotlib
+
+import sphinx_rtd_theme
+
 
 # Use matplotlib agg backend
 matplotlib.use("agg")
@@ -25,15 +26,15 @@ matplotlib.use("agg")
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-sys.path.append(os.path.abspath('_ext'))
+sys.path.insert(0, os.path.abspath('.') + os.sep + '_ext')
 
 
 # -- Project information -----------------------------------------------------
 project = 'ObsPy'
 author = 'The ObsPy Development Team (devs@obspy.org)'
 year = datetime.date.today().year
-copyright = '2012-{}, The ObsPy Development Team (devs@obspy.org)'.format(year)
-version = ".".join(obspy.__version__.split(".")[:3])
+copyright = '%d, The ObsPy Development Team (devs@obspy.org)' % (year)
+version = ".".join(obspy.__version__.split(".")[:2])
 release = obspy.__version__
 
 
@@ -49,58 +50,34 @@ extensions = [
     'sphinx.ext.intersphinx',
     'sphinx.ext.doctest',
     'sphinx.ext.autodoc',
+    'sphinx.ext.ifconfig',
     'sphinx.ext.viewcode',
     'sphinx.ext.mathjax',
     'sphinx.ext.autosummary',
+    'sphinx.ext.graphviz',
+    'sphinx.ext.inheritance_diagram',
     'sphinx.ext.coverage',
     'sphinx.ext.duration',
     'sphinx.ext.todo',
     # theme
-    'm2r2',
     'sphinx_rtd_theme',
     # custom extensions
     'plot_directive',
     'credits',
     'citations',
-    # 'sphinxcontrib.bibtex',
 ]
-
-# Uncomment this if you want to use sphinxcontrib.bibtex
-# from pathlib import Path
-# bibtex_bibfiles = list(Path('.').joinpath('source', 'bibliography').glob('*.bib'))
-
-# The file extensions of source files. Sphinx considers the files with this
-# suffix as sources. The value can be a dictionary mapping file extensions to
-# file types.
-source_suffix = {
-    '.rst': 'restructuredtext',
-    '.md': 'markdown',
-}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
-# The master toctree document.
-master_doc = 'index'
-
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = [
-    '_ext', '_images', '_static', '_templates', 'bibliography', 'credits']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
 # Warn about all references where the target cannot be found.
 nitpicky = True
-nitpick_ignore = [
-    ('py:class', 'list of str'),
-    ('py:class', 'optional'),
-    ('py:class', 'file'),
-]
-
-# suppress built-in types by default in nitpick
-import builtins
-for name in dir(builtins):
-    nitpick_ignore += [('py:class', name)]
+nitpick_ignore = ['norm_resp']
 
 # configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
@@ -109,27 +86,17 @@ intersphinx_mapping = {
     'scipy': ('https://docs.scipy.org/doc/scipy/reference/', None),
     'matplotlib': ('https://matplotlib.org/', None),
     'sqlalchemy': ('https://docs.sqlalchemy.org/en/latest/', None),
[…]
