+++ misc/docs/source/_ext/citations.py
@@ -2,7 +2,7 @@
 
 import glob
 import os
-from distutils.version import LooseVersion
+import re
 
 import pybtex
 from pybtex.database.input import bibtex
@@ -10,28 +10,41 @@ from pybtex.style.names.lastfirst import NameStyle
 from pybtex.style.template import field, join, node, optional, sentence, words
 
 
+TEMPLATE = """
+.. _citations:
+
+.. DON'T EDIT THIS FILE MANUALLY!
+   Instead insert a BibTeX file into the bibliography folder.
+   This file will be recreated automatically during building the docs.
+
+Citations
+==========
+
+.. list-table::
+   :widths: 1 4
+
+"""
+
+
 REPLACE_TOKEN = [
-    (u"<nbsp>", u" "),
-    (u"\xa0", u" "),
-    (u'–', '-'),
-    (u'—', '-'),
-    (u'—', '-'),
-    (u'--', '-'),
-    (u"\'{a}", u"á"),
-    (u"{\\ae}", u"æ"),
-    (u"**{", u"**"),
-    (u"}**", u"**"),
-    (u"}**", u"**"),
+    ("<nbsp>", " "),
+    ("\xa0", " "),
+    ('–', '-'),
+    ('—', '-'),
+    ('—', '-'),
+    ('--', '-'),
+    ("\'{a}", "á"),
+    ("{\\ae}", "æ"),
+    ("**{", "**"),
+    ("}**", "**"),
+    ("}**", "**"),
 ]
 
 
 @node
 def names(children, data, role, **kwargs):
     assert not children
-    if LooseVersion(pybtex.__version__) >= '0.22.0':
-        persons = data['entry'].persons[role]
-    else:
-        persons = data.persons[role]
+    persons = data['entry'].persons[role]
     return join(**kwargs)[[NameStyle().format(person, abbr=True)
                            for person in persons]].format_data(data)
 
@@ -120,54 +133,69 @@ formats = {
     ],
 }
 
-parser = bibtex.Parser(encoding='utf8')
+def create_citations_page(app):
+    parser = bibtex.Parser(encoding='utf8')
+    for file in glob.glob(os.path.join('source', 'bibliography', '*.bib')):
+        try:
+            parser.parse_file(file)
+        except:
+            print("Error parsing file %s:" % (file))
+            raise
 
-for file in glob.glob(os.path.join('source', 'bibliography', '*.bib')):
-    try:
-        parser.parse_file(file)
-    except Exception:
-        print("Error parsing file %s:" % (file))
-        raise
+    entries = parser.data.entries
 
-entries = parser.data.entries
+    # create citations.rst
+    fh = open(os.path.join('source', 'citations.rst'),
+              mode='w', encoding='utf-8', newline='\n')
+    fh.write(TEMPLATE)
 
-# write index.rst
-fh = open(os.path.join('source', 'citations.rst'), 'wb')
-fh.write(b"""
-.. _citations:
-
-.. DON'T EDIT THIS FILE MANUALLY!
-   Instead insert a BibTeX file into the bibliography folder.
-   This file will be recreated automatically during building the docs.
-   (The creation of the file can be tested manually by running
-    ``make citations`` from command line)
-
-Citations
-==========
-
-.. list-table::
-   :widths: 1 4
-
-""")
-
-for key in sorted(entries.keys()):
-    entry = entries[key]
-    if entry.type not in formats:
-        msg = "BibTeX entry type %s not implemented"
-        raise NotImplementedError(msg % (entry.type))
-    out = '   * - .. [%s]%s'
-    if LooseVersion(pybtex.__version__) >= '0.22.0':
+    for key in sorted(entries.keys()):
+        entry = entries[key]
+        if entry.type not in formats:
+            msg = "BibTeX entry type %s not implemented"
[…]
