+++ misc/docs/source/_templates/script.rst
@@ -1,11 +1,0 @@
-{{ fullname }}
-{{ underline }}
-
-.. note::
-    This script automatically installs during the setup procedure with the
-    name ``$ obspy-{{ objname }}``. For more info on the command line options,
-    please run ``$ obspy-{{ objname }} --help``.
-    Alternatively you can also execute ``$ python -m {{ fullname }}``.
-
-.. currentmodule:: {{ fullname }}
-.. automodule:: {{ fullname }}
