+++ misc/docs/DEPENDENCIES.txt
@@ -1,0 +1,9 @@
+sphinx>=3.1.1
+sphinx_rtd_theme
+pybtex>=0.22.2
+obspy
+matplotlib
+
+# tutorial
+basemap
+mlpy
