+++ misc/docs/source/conf.py
@@ -8,6 +8,7 @@
 #
 
 import datetime
+import glob
 import obspy
 import os
 import sys
@@ -26,15 +27,15 @@ matplotlib.use("agg")
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-sys.path.insert(0, os.path.abspath('.') + os.sep + '_ext')
+sys.path.append(os.path.abspath('_ext'))
 
 
 # -- Project information -----------------------------------------------------
 project = 'ObsPy'
 author = 'The ObsPy Development Team (devs@obspy.org)'
 year = datetime.date.today().year
-copyright = '%d, The ObsPy Development Team (devs@obspy.org)' % (year)
-version = ".".join(obspy.__version__.split(".")[:2])
+copyright = '2012-{}, The ObsPy Development Team (devs@obspy.org)'.format(year)
+version = ".".join(obspy.__version__.split(".")[:3])
 release = obspy.__version__
 
 
@@ -50,16 +51,14 @@ extensions = [
     'sphinx.ext.intersphinx',
     'sphinx.ext.doctest',
     'sphinx.ext.autodoc',
-    'sphinx.ext.ifconfig',
     'sphinx.ext.viewcode',
     'sphinx.ext.mathjax',
     'sphinx.ext.autosummary',
-    'sphinx.ext.graphviz',
-    'sphinx.ext.inheritance_diagram',
     'sphinx.ext.coverage',
     'sphinx.ext.duration',
     'sphinx.ext.todo',
     # theme
+    'm2r',
     'sphinx_rtd_theme',
     # custom extensions
     'plot_directive',
@@ -67,17 +66,38 @@ extensions = [
     'citations',
 ]
 
+# The file extensions of source files. Sphinx considers the files with this
+# suffix as sources. The value can be a dictionary mapping file extensions to
+# file types.
+source_suffix = {
+    '.rst': 'restructuredtext',
+    '.md': 'markdown',
+}
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
+
+# The master toctree document.
+master_doc = 'index'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = [
+    '_ext', '_images', '_static', '_templates', 'bibliography', 'credits']
 
 # Warn about all references where the target cannot be found.
 nitpicky = True
-nitpick_ignore = ['norm_resp']
+nitpick_ignore = [
+    ('py:class', 'list of str'),
+    ('py:class', 'optional'),
+    ('py:class', 'file'),
+]
+
+# suppress built-in types by default in nitpick
+import builtins
+for name in dir(builtins):
+    nitpick_ignore += [('py:class', name)]
 
 # configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
@@ -86,17 +106,27 @@ intersphinx_mapping = {
     'scipy': ('https://docs.scipy.org/doc/scipy/reference/', None),
     'matplotlib': ('https://matplotlib.org/', None),
     'sqlalchemy': ('https://docs.sqlalchemy.org/en/latest/', None),
+    'pip': ('https://pip.pypa.io/en/stable/', None),
 }
+
+# A boolean that decides whether module names are prepended to all object names
+# (for object types where a “module” of some kind is defined).
+add_module_names = False
+
+# A list of ignored prefixes for module index sorting.
+modindex_common_prefix = ['obspy.']
+
+# These values determine how to format the current date.
[…]
