+++ misc/docs/source/_ext/credits.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+"""
+Directive for creating credits.rst page listing contributors, funds and quotes.
+"""
 
 import os
 
@@ -15,22 +18,25 @@ it is. These people have helped by writing code and documentation, and by
 testing. They have created and maintained this product, its associated
 libraries and applications, our build tools and our web sites.
 
-.. rubric:: Hall of Fame
+Hall of Fame
+------------
 
 .. hlist::
-    :columns: 3
+    :columns: 2
 
 """
 
 TEMPLATE_FUNDS = """
-.. rubric:: Funds
+Funds
+-----
 
 ObsPy was partially funded by the
 
 """
 
 TEMPLATE_QUOTES = """
-.. rubric:: Quotes
+Quotes
+------
 
 """
 
@@ -78,9 +84,11 @@ def create_credits_page(app):
 
     for item in funds:
         item = item.split('---')
-        fh.write("\nepigraph::\n    %s" % (item[0]))
+        # quote
+        fh.write("\n\n    %s" % (item[0].strip()))
+        # attribution, if given
         try:
-            fh.write("\n\n-- %s" % (item[1]))
+            fh.write("\n\n    -- %s" % (item[1]))
         except IndexError:
             pass
     fh.close()
