+++ misc/docs/source/contents.rst
@@ -1,0 +1,17 @@
+.. _contents:
+
+Table of Contents
+=================
+
+.. toctree::
+   :maxdepth: 3
+
+   tutorial/index
+   gallery
+   packages/index
+   changelog
+   citations
+   coding_style
+   contributing
+   code_of_conduct
+   credits
