+++ misc/docs/DEPENDENCIES.txt
@@ -1,9 +1,10 @@
 sphinx>=3.1.1
-sphinx_rtd_theme
+m2r==dev  # due to https://github.com/miyakogi/m2r/issues/51
 pybtex>=0.22.2
-obspy
+sphinx_rtd_theme
 matplotlib
+obspy
 
-# tutorial
-basemap
-mlpy
+# deps used in tutorial but not working anymore
+# basemap
+# mlpy

[…]
