### 2020-06-25 4762af2
FDSN: set timeout according to default or user setting in service discovery
