+++ CHANGELOG.txt
@@ -11,6 +11,7 @@ Changes:
      station (see #2611)
    * update RASPISHAKE URL mapping to use https
    * fix a bug of not handling HTTPException in mass_downloader (see #2606)
+   * use the client's set timeout in service discovery, too (see #2656)
  - obspy.clients.filesystem:
    * sds: continue get_all_stations() even if encountering an invalid channel
      code (see #2636)
