### 2020-06-25 eb2ffb6
I added a test case for stage 1 having no poles or zeros.
Flake8 formatting fix.

Switched to a smaller test file updated the CHANGELOG.

remove big test file again, and squash commits
