### 2020-06-26 a318189
pha: deal with nan magnitude, add test
