### 2020-06-26 2450e66
fdsn: fix setting timeout on the threaded service discovery
