### 2020-06-26 0d68b74
PPSD plotting: adjust to matplotlib API change for set_clim()
