### 2020-06-26 73b4937
fdsn: add url mapping for uib-norsar
