### 2020-06-26 ebee907
pha: move new tests to own method
