### 2020-06-26 cdc3658
pha: be less strict in detection of pha format
