+++ CHANGELOG.txt
@@ -1,4 +1,4 @@
-1.2.2rc1
+1.2.2rc2
 ========
 
 Changes:
