+++ CHANGELOG.txt
@@ -1,4 +1,4 @@
-1.2.2rc2
+1.2.2rc3
 ========
 
 Changes:
