+++ CHANGELOG.txt
@@ -1,5 +1,5 @@
-1.2.2rc3
-========
+1.2.2 (doi: 10.5281/zenodo.3921997)
+===================================
 
 Changes:
  - obspy.core:
