### 2020-06-30 04c80df
remove obspy.clients.arclink and all other occurences (docs etc.)
Arclink has been switched off by all big datacenters, so actual use
cases for it will be very minor if at all present. legacy applications
can use older obspy versions.
