+++ obspy/clients/arclink/client.py
@@ -1,1307 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-ArcLink/WebDC client for ObsPy.
-
-:copyright:
-    The ObsPy Development Team (devs@obspy.org)
-:license:
-    GNU Lesser General Public License, Version 3
-    (https://www.gnu.org/copyleft/lesser.html)
-"""
-import io
-import os
-import time
-import warnings
-from fnmatch import fnmatch
-from telnetlib import Telnet
-
-from lxml import etree, objectify
-import numpy as np
-
-from obspy import UTCDateTime, read
-from obspy.core.util import AttribDict, complexify_string
-from obspy.core.util.deprecation_helpers import ObsPyDeprecationWarning
-
-
-DCID_KEY_FILE = os.path.join(os.getenv('HOME') or '', 'dcidpasswords.txt')
-MAX_REQUESTS = 50
-
-_ROUTING_NS_1_0 = "http://geofon.gfz-potsdam.de/ns/Routing/1.0/"
-_ROUTING_NS_0_1 = "http://geofon.gfz-potsdam.de/ns/routing/0.1/"
-_INVENTORY_NS_1_0 = "http://geofon.gfz-potsdam.de/ns/Inventory/1.0/"
-_INVENTORY_NS_0_2 = "http://geofon.gfz-potsdam.de/ns/inventory/0.2/"
-
-MSG_NOPAZ = "No Poles and Zeros information returned by server."
-
-
-class ArcLinkException(Exception):
-    """
-    Raised by the ArcLink client for known exceptions.
-    """
-
-
-class Client(object):
-    """
-    DEPRECATED -- ArcLink protocol is officially deprecated
-
-    The ArcLink/WebDC client.
-
-    :type user: str
-    :param user: The user name is used for identification with the ArcLink
-        server. This entry is also used for usage statistics within the data
-        centers, so please provide a meaningful user id such as your email
-        address.
-    :type host: str, optional
-    :param host: Host name of the remote ArcLink server (default host is
-        ``'webdc.eu'``).
-    :type port: int, optional
-    :param port: Port of the remote ArcLink server (default port is ``18002``).
-    :type timeout: int, optional
-    :param timeout: Seconds before a connection timeout is raised (default is
-        ``20`` seconds).
-    :type password: str, optional
-    :param password: A password used for authentication with the ArcLink server
-        (default is an empty string).
-    :type institution: str, optional
-    :param institution: A string containing the name of the institution of the
-        requesting person (default is an ``'Anonymous'``).
-    :type dcid_keys: dict, optional
-    :param dcid_keys: Dictionary of data center ids (DCID) and passwords used
-        for decoding encrypted waveform requests.
-    :type dcid_key_file: str, optional
-    :param dcid_key_file: Simple text configuration file containing lines of
-        data center ids (DCIDs) and password pairs separated by a equal sign,
-        e.g. for DCID ``BIA`` and password ``OfH9ekhi`` use ``"BIA=OfH9ekhi"``.
-        If not set, passwords found in a file called `$HOME/dcidpasswords.txt`
-        will be used automatically.
-    :type debug: bool, optional
-    :param debug: Enables verbose output of the connection handling (default is
-        ``False``).
-    :type command_delay: float, optional
-    :param command_delay: Delay between each command send to the ArcLink server
-        (default is ``0``).
-    :type status_delay: float, optional
-    :param status_delay: Delay in seconds between each status request (default
-        is ``0.5`` seconds).
-    """
-    max_status_requests = MAX_REQUESTS
-
-    def __init__(self, user, host="webdc.eu", port=18002,
-                 password="", institution="Anonymous", timeout=20,
-                 dcid_keys={}, dcid_key_file=None, debug=False,
[…]
