+++ obspy/clients/fdsn/tests/test_client.py
@@ -119,7 +119,7 @@ class ClientTestCase(unittest.TestCase):
 
         test_urls_valid = list(URL_MAPPINGS.values())
         test_urls_valid += [
-            "http://arclink.ethz.ch",
+            "http://something.ethz.ch",
             "http://example.org",
             "https://webservices.rm.ingv.it",
             "http://localhost:8080/test/",
