+++ obspy/core/util/base.py
@@ -41,7 +41,7 @@ DEFAULT_MODULES = ['clients.filesystem', 'core', 'db', 'geodetics', 'imaging',
                    'io.seiscomp', 'io.stationtxt', 'io.stationxml', 'io.wav',
                    'io.win', 'io.xseed', 'io.y', 'io.zmap', 'realtime',
                    'scripts', 'signal', 'taup']
-NETWORK_MODULES = ['clients.arclink', 'clients.earthworm', 'clients.fdsn',
+NETWORK_MODULES = ['clients.earthworm', 'clients.fdsn',
                    'clients.iris', 'clients.neic', 'clients.nrl',
                    'clients.seedlink', 'clients.seishub', 'clients.syngine']
 ALL_MODULES = DEFAULT_MODULES + NETWORK_MODULES
