+++ misc/docs/source/packages/index.rst
@@ -55,7 +55,6 @@ categories.*
     :toctree: .
     :nosignatures:
 
-    obspy.clients.arclink
     obspy.clients.earthworm
     obspy.clients.fdsn
     obspy.clients.filesystem
