### 2020-07-01 b44b8e2
Removes further matplotlib version checks
Signed-off-by: Damian Kula <dkula@unistra.fr>
