### 2020-07-01 b6d09c0
Removes compatibility checks with matplotlib <3
Signed-off-by: Damian Kula <dkula@unistra.fr>
