+++ obspy/core/tests/test_channel.py
@@ -42,11 +42,7 @@ class ChannelTestCase(unittest.TestCase):
         """
         Tests the response plot.
         """
-        # Bug in matplotlib 1.4.0 - 1.4.x:
-        # See https://github.com/matplotlib/matplotlib/issues/4012
         reltol = 1.0
-        if [1, 4, 0] <= MATPLOTLIB_VERSION <= [1, 5, 0]:
-            reltol = 2.0
 
         cha = read_inventory()[0][0][0]
         with warnings.catch_warnings(record=True):
