+++ obspy/core/tests/test_response.py
@@ -126,11 +126,7 @@ class ResponseTestCase(unittest.TestCase):
         """
         Tests the response plot.
         """
-        # Bug in matplotlib 1.4.0 - 1.4.x:
-        # See https://github.com/matplotlib/matplotlib/issues/4012
         reltol = 1.0
-        if [1, 4, 0] <= MATPLOTLIB_VERSION <= [1, 5, 0]:
-            reltol = 2.0
 
         resp = read_inventory()[0][0][0].response
         with warnings.catch_warnings(record=True):
@@ -145,11 +141,7 @@ class ResponseTestCase(unittest.TestCase):
         """
         Tests the response plot in degrees.
         """
-        # Bug in matplotlib 1.4.0 - 1.4.x:
-        # See https://github.com/matplotlib/matplotlib/issues/4012
         reltol = 1.0
-        if [1, 4, 0] <= MATPLOTLIB_VERSION <= [1, 5, 0]:
-            reltol = 2.0
 
         resp = read_inventory()[0][0][0].response
         with warnings.catch_warnings(record=True):
