+++ obspy/imaging/tests/test_ppsd.py
@@ -43,11 +43,7 @@ class PPSDTestCase(unittest.TestCase):
         """
         Test plot of ppsd example data, normal (non-cumulative) style.
         """
-        # mpl < 2.2 has slightly offset ticks/ticklabels, so needs a higher
-        # tolerance (see e.g. http://tests.obspy.org/102260)
         reltol = 2
-        if MATPLOTLIB_VERSION < [2, 2]:
-            reltol = 4
         # Catch underflow warnings due to plotting on log-scale.
         with np.errstate(all='ignore'):
             with ImageComparison(self.path, 'ppsd_freq.png',
