### 2020-07-01 eb060e4
Removes unnecessary MATPLOTLIB_VERSION imports
