+++ obspy/core/tests/test_station.py
@@ -17,7 +17,6 @@ import numpy as np
 from matplotlib import rcParams
 
 from obspy import read_inventory, UTCDateTime
-from obspy.core.util import MATPLOTLIB_VERSION
 from obspy.core.util.testing import ImageComparison
 
 
