### 2020-07-01 4b37398
point to proper obspy_github_api package
