### 2020-07-02 89da4f8
use polyval/polyfit from numpy, deprecated and will be removed in scipy 2.0
fixes #2666
