### 2020-07-02 75c6572
should fixes flake8 issues for io.xseed
unfortunatly I can't find a single RESP example containing blockette 48
or 59 *with* any calibration data - therefore no test case atm ...
