### 2020-07-02 2d6862e
Introduce fine-grained FDSN client exceptions and tests
