+++ obspy/clients/fdsn/client.py
@@ -29,7 +29,11 @@ from obspy.core.compatibility import collections_abc
 from .header import (DEFAULT_PARAMETERS, DEFAULT_USER_AGENT, FDSNWS,
                      OPTIONAL_PARAMETERS, PARAMETER_ALIASES, URL_MAPPINGS,
                      WADL_PARAMETERS_NOT_TO_BE_PARSED, DEFAULT_SERVICES,
-                     FDSNException, FDSNRedirectException, FDSNNoDataException)
+                     FDSNException, FDSNRedirectException, FDSNNoDataException,
+                     FDSNTimeoutException, FDSNAuthenticationException,
+                     FDSNBadRequestException, FDSNNoServiceException,
+                     FDSNServerException, FDSNTooManyRequestsException,
+                     FDSNRequestTooLargeException)
 from .wadl_parser import WADLParser
 
 from urllib.parse import urlencode
@@ -344,7 +348,7 @@ class Client(object):
             msg = ("EIDA token authentication requested but service at '{}' "
                    "does not specify /dataselect/auth in the "
                    "dataselect/application.wadl.").format(self.base_url)
-            raise FDSNException(msg)
+            raise FDSNAuthenticationException(msg)
 
         token_file = None
         # check if there's a local file that matches the provided string
@@ -1494,7 +1498,8 @@ class Client(object):
                 redirect_messages.add(str(wadl))
                 continue
             elif decoded_wadl == "timeout":
-                raise FDSNException("Timeout while requesting '%s'." % url)
+                raise FDSNTimeoutException("Timeout while requesting '%s'."
+                                           % url)
 
             if "dataselect" in url:
                 wadl_parser = WADLParser(wadl)
@@ -1537,7 +1542,7 @@ class Client(object):
             msg = ("No FDSN services could be discovered at '%s'. This could "
                    "be due to a temporary service outage or an invalid FDSN "
                    "service address." % self.base_url)
-            raise FDSNException(msg)
+            raise FDSNNoServiceException(msg)
 
         # Cache.
         if self.debug is True:
@@ -1711,18 +1716,24 @@ def raise_on_error(code, data):
         raise FDSNNoDataException("No data available for request.",
                                   server_info)
     elif code == 400:
-        msg = ("Bad request. If you think your request was valid "
-               "please contact the developers.")
-        raise FDSNException(msg, server_info)
+        if "token is expired" in server_info:
+            raise FDSNAuthenticationException('Token is expired',
+                                              server_info)
+        else:
+            msg = ("Bad request. If you think your request was valid "
+                   "please contact the developers.")
+            raise FDSNBadRequestException(msg, server_info)
     elif code == 401:
-        raise FDSNException("Unauthorized, authentication required.",
-                            server_info)
+        raise FDSNAuthenticationException("Unauthorized, authentication "
+                                          "required.", server_info)
     elif code == 403:
-        raise FDSNException("Authentication failed.", server_info)
+        raise FDSNAuthenticationException("Authentication failed.",
+                                          server_info)
     elif code == 413:
-        raise FDSNException("Request would result in too much data. "
-                            "Denied by the datacenter. Split the request "
-                            "in smaller parts", server_info)
+        raise FDSNRequestTooLargeException("Request would result in too much "
+                                           "data. Denied by the datacenter. "
+                                           "Split the request in smaller "
+                                           "parts", server_info)
     # Request URI too large.
     elif code == 414:
         msg = ("The request URI is too large. Please contact the ObsPy "
@@ -1731,15 +1742,16 @@ def raise_on_error(code, data):
     elif code == 429:
[…]
