+++ CHANGELOG.txt
@@ -7,6 +7,8 @@ Changes:
  - obspy.io.xseed:
    * fix a bug reading SEED blockettes 48 and 58 which was likely never
      encountered (see #2668)
+ - obspy.clients.fdsn:
+   * introduce fine-grained FDSN client exceptions (see #2653)
 
 1.2.2 (doi: 10.5281/zenodo.3921997)
 ===================================
