+++ obspy/CONTRIBUTORS.txt
@@ -69,6 +69,7 @@ Meschede, Matthias
 Michelini, Alberto
 Miller, Nathaniel C.
 Morgenstern, Bernhard
+Münchmeyer, Jannes
 Murray-Bergquist, Louisa
 Nesbitt, Ian M.
 Nof, Ran Novitsky
