### 2020-07-02 247e40b
fix small flake8 issue
