### 2020-07-03 cd19f3a
Addresses review comments
Signed-off-by: Damian Kula <dkula@unistra.fr>
