### 2020-07-03 9799a34
Raise FDSNAuthenticationException if token and user/password are provided
