### 2020-07-03 21c4276
Extend FDSN exception tests and use mocks for the tests
