### 2020-07-06 e14834e
updates to core.py and packet.py to deal with mass position channels (sampling_rate=0.1Hz and 16 encoding)
