### 2020-07-06 2a61936
Add status code to FDSN exceptions and introduce more exceptions to ensure 1:1 mapping between exceptions and HTTP status codes
