+++ .github/workflows/runtests.yml
@@ -8,7 +8,7 @@ env:
   CI_URL: https://github.com/obspy/obspy/actions/runs/${{ github.run_id }}
   CONFIG_PATH: obspy_ci_config/conf.json
   GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-  OBSPY_GITHUB_BRANCH: release_0.9.0
+  OBSPY_GITHUB_BRANCH: gh-actions
 
 
 jobs:
