### 2020-07-08 747c0e5
sds report: add optional checks for voltage and gps
currently only parses voltage from *.*.*.VEI channel and number of
satellites from *.*.*.GNS channel (as done by Centaur) but could be
extended to how other digitizers store that info if needed.
