+++ obspy/scripts/sds_html_report.py
@@ -24,6 +24,9 @@ For a subsequent update of latency only:
     $ obspy-sds-report -r /bay200/mseed_online/archive -o /tmp/sds_report \
 --update
 
+New options have been added, to also check state-of-health files for input
+power voltage (channel `VEI`) and number of GPS satellites (channel `GNS`).
+
 Screenshot of resulting html page (cut off at the bottom):
 
 .. image:: /_images/sds_report.png
@@ -56,18 +59,18 @@ p.monospace_wide_columns
      -webkit-column-count: 4;
      -moz-column-count: 4;
      column-count: 4;
-     -webkit-column-width: 480px;
-     -moz-column-width: 480px;
-     column-width: 480px;
+     -webkit-column-width: {column_width_wide}px;
+     -moz-column-width: {column_width_wide}px;
+     column-width: {column_width_wide}px;
      }}
 p.monospace_narrow_columns
     {{font-family: "Lucida Console", Monaco, monospace;
-     -webkit-column-count: 8;
-     -moz-column-count: 8;
-     column-count: 8;
-     -webkit-column-width: 250px;
-     -moz-column-width: 250px;
-     column-width: 250px;
+     -webkit-column-count: 10;
+     -moz-column-count: 10;
+     column-count: 10;
+     -webkit-column-width: {column_width_narrow}px;
+     -moz-column-width: {column_width_narrow}px;
+     column-width: {column_width_narrow}px;
      }}
 </style>
 </head>
@@ -92,7 +95,7 @@ p.monospace_narrow_columns
   <span style="background-color: {latency_error_color}">
       {latency_error_info:*<40s}</span><br>
   <span style="background-color: {outdated_color}">
-      {outdated_info:*<40s}</span><br>
+      {outdated_info:*<40s}</span><br>{legend_volt}{legend_gps}
 </p>
   <img src="{output_basename}.png" alt="obspy-scan image broken!">
 </body>
@@ -101,7 +104,14 @@ p.monospace_narrow_columns
 HTML_LINE = (
     '<span style="background-color: {color}">'
     '{network:*<2s} {station:*<6s} {location:*<2s} {channel:*<3s} '
-    '{latency_string} {percentage_string} {gap_count_string}</span><br>\n')
+    '{latency_string} {percentage_string} {gap_count_string} {volt_string} '
+    '{gps_count_string}</span><br>\n')
+HTML_LEGEND_VOLT = """
+  <span style="background-color: {volt_warn_color}">
+      {volt_warn_info:*<40s}</span><br>"""
+HTML_LEGEND_GPS = """
+  <span style="background-color: {gps_warn_color}">
+      {gps_warn_info:*<40s}</span><br>"""
 
 
 def _latency_to_tuple(latency):
@@ -154,12 +164,13 @@ def _latency_info_string(latency, only_days=False, pad=True):
 
 
 def _latency_line_html(latency_tuple, args, color=None, only_days=False,
-                       gap_info=True):
+                       gap_info=True, volt_info=True, gps_info=True):
     """
     Format a single latency information tuple (net, sta, loc, cha, latency,
     percentage, gap count) as a html line.
     """
-    net, sta, loc, cha, latency, percentage, gap_count = latency_tuple
+    net, sta, loc, cha, latency, percentage, gap_count, volt, gps_count = \
+        latency_tuple
     if not color:
         if (latency is None or np.isinf(latency) or
                 latency > args.check_back_days * 24 * 3600):
@@ -168,11 +179,17 @@ def _latency_line_html(latency_tuple, args, color=None, only_days=False,
             color = args.latency_error_color
         elif latency > args.latency_warn:
             color = args.latency_warn_color
[…]
