### 2020-07-08 f6078a8
fixes PEP8 variables too long
