### 2020-07-08 42b94dd
Updates to string representations of Reftek object and Packets - sampling_rate is now a float
