+++ obspy/io/reftek/__init__.py
@@ -94,7 +94,7 @@ Packet Sequence  Byte Count  Data Fmt  Sampling Rate      Time
   | Packet Type   |  Event #  | Station | Channel #         |
   |   |  Unit ID  |    | Data Stream #  |   |  # of samples |
   |   |   |  Exper.#   |   |  |  |      |   |    |          |
-0000 EH AE4C  0  416  427  0 C0 KW1    200         2015-10-09T22:50:51.000000Z
+0000 EH AE4C  0  416  427  0 C0 KW1   200.0        2015-10-09T22:50:51.000000Z
 0001 DT AE4C  0 1024  427  0 C0             0  549 2015-10-09T22:50:51.000000Z
 0002 DT AE4C  0 1024  427  0 C0             1  447 2015-10-09T22:50:51.000000Z
 0003 DT AE4C  0 1024  427  0 C0             2  805 2015-10-09T22:50:51.000000Z
@@ -122,7 +122,7 @@ Packet Sequence  Byte Count  Data Fmt  Sampling Rate      Time
 0025 DT AE4C  0 1024  427  0 C0             1  673 2015-10-09T22:51:22.025000Z
 0026 DT AE4C  0 1024  427  0 C0             2  759 2015-10-09T22:51:21.595000Z
 0027 DT AE4C  0 1024  427  0 C0             0   67 2015-10-09T22:51:25.055000Z
-0028 ET AE4C  0  416  427  0 C0 KW1    200         2015-10-09T22:50:51.000000Z
+0028 ET AE4C  0  416  427  0 C0 KW1   200.0        2015-10-09T22:50:51.000000Z
 (detailed packet information with: 'print(Reftek130.__str__(compact=False))')
 >>> print(rt.__str__(compact=False)) # doctest: +NORMALIZE_WHITESPACE +ELLIPSIS
 Reftek130 (29 packets, file: ...225051000_00008656)
@@ -146,7 +146,7 @@ EH Packet
     last_sample_time: None
     position:
     ...
-    sampling_rate: 200
+    sampling_rate: 200.0
     station_channel_number: (None, None, None, None, None, None, None, ...)
     station_comment: STATION COMMENT
     station_name: KW1
