### 2020-07-08 441232a
Updates to __str__() method in EHPacket class to account for new string representations of Reftek object and Packets
