### 2020-07-09 83a9b67
Add interpolation-option for all stage types
