### 2020-07-10 62dc623
Removes usages of reltol argument if it's not changed from 1 which is default
Signed-off-by: Damian Kula <dkula@unistra.fr>
