### 2020-07-10 63635d8
Updated changelog and contributors files, fixed width of string representations, and changed bytes_ to be static
