+++ obspy/io/reftek/packet.py
@@ -154,12 +154,12 @@ class EHPacket(Packet):
             sta = (self.station_name.strip() +
                    self.station_name_extension.strip())
             info = ("{:04d} {:2s} {:4s} {:2d} {:4d} {:4d} {:2d} {:2s} "
-                    "{:5s} {:4.1f}        {!s}").format(
+                    "{:5s}  {:4s}        {!s}").format(
                         self.packet_sequence, self.type.decode(),
                         self.unit_id.decode(), self.experiment_number,
                         self.byte_count, self.event_number,
                         self.data_stream_number, self.data_format.decode(),
-                        sta, self.sampling_rate, self.time)
+                        sta, str(self.sampling_rate)[:4], self.time)
         else:
             info = []
             for key in self._headers:
