### 2020-07-10 dfb9a3a
Updates to changelog and contibutors files
Update name ordering in contributors file
