+++ obspy/CONTRIBUTORS.txt
@@ -77,6 +77,7 @@ Panning, Mark P.
 Rapagnani, Giovanni
 Parker, Tom
 Pestourie, Romain
+Pourpoint, Maeva
 Reyes, Celso
 Ringler, Adam
 Rothenhäusler, Nicolas
