### 2020-07-11 cd45a81
Update obspy/core/inventory/response.py
Co-authored-by: Tobias Megies <megies@users.noreply.github.com>
