+++ obspy/core/inventory/response.py
@@ -1623,7 +1623,7 @@ class Response(ComparingObject):
             used (disregarding all later stages).
         :type hide_sensitivity_mismatch_warning: bool
         :param hide_sensitivity_mismatch_warning: Hide the evalresp warning
-            that computed and reported sensitivities don't match.
+            that computed and reported sensitivities do not match.
         :rtype: :class:`numpy.ndarray`
         :returns: frequency response at requested frequencies
         """
