### 2020-07-13 2ddbbdd
Support for volts and temperature.
