### 2020-07-14 dec48a2
hint at option to hide sensitivity warning
