### 2020-07-14 fce7cc1
add documentation-strings
