+++ obspy/core/inventory/response.py
@@ -340,6 +340,8 @@ class PolesZerosResponseStage(ResponseStage):
         Produce the response curve from this stage's data for a given
         range of frequencies
         :param frequencies: Frequency range to get resp curve over
+        :param fast: Indicates whether to speed up calculation through
+            interpolation.
         :return: The curve describing this response stage
         """
         # Has to be imported here for now to avoid circular imports.
@@ -361,9 +363,9 @@ class PolesZerosResponseStage(ResponseStage):
             amp = np.abs(resp)
             phase = np.radians(np.unwrap(np.angle(resp, deg=False))) / np.pi
             amp = scipy.interpolate.InterpolatedUnivariateSpline(
-                    resp_frequencies, amp, k=2)(frequencies)
+                resp_frequencies, amp, k=2)(frequencies)
             phase = scipy.interpolate.InterpolatedUnivariateSpline(
-                    resp_frequencies, phase, k=2)(frequencies)
+                resp_frequencies, phase, k=2)(frequencies)
             final_resp = np.zeros_like(frequencies) + 0j
             final_resp.real = amp * np.cos(phase)
             final_resp.imag = amp * np.sin(phase)
@@ -547,6 +549,8 @@ class CoefficientsTypeResponseStage(ResponseStage):
         Produce the response curve from this coefficient
         response stage for a range of frequencies
         :param frequencies: Frequency range to get resp curve over
+        :param fast: Indicates whether to speed up calculation through
+            interpolation.
         :return: The curve describing this response stage
         """
         # Decimation blockette, e.g. gain only!
@@ -580,7 +584,9 @@ class CoefficientsTypeResponseStage(ResponseStage):
                 # we get the numerator and denominator and do the math
                 # on them in their representation as magnitude and
                 # phase rather than standard complex format
-                w = resp_frequencies  # rename to be concise and match conventions
+                
+                # rename to be concise and match conventions
+                w = resp_frequencies
 
                 resp = np.zeros_like(w) + 0j
                 for idx, num in enumerate(self.numerator):
@@ -810,7 +816,10 @@ class FIRResponseStage(ResponseStage):
 
     def get_response(self, frequencies, fast=True):
         """
-        Given Computes the 
+        Given Computes the
+        :param frequencies: Discrete frequencies to calculate response for. 
+        :param fast: Indicates whether to speed up calculation through
+            interpolation.
         """
         # Decimation blockette, e.g. gain only!
         if not len(self._coefficients):
