### 2020-07-14 6054a9f
flake8-formatting fixes
