### 2020-07-14 0461068
Add fast-option to remove_response
