+++ obspy/core/trace.py
@@ -2667,7 +2667,7 @@ seismometer_correction_simulation.html#using-a-resp-file>`_.
 
     @_add_processing_info
     def remove_response(self, inventory=None, output="VEL", water_level=60,
-                        pre_filt=None, zero_mean=True, taper=True,
+                        pre_filt=None, zero_mean=True, fast=True, taper=True,
                         taper_fraction=0.05, plot=False, fig=None, **kwargs):
         """
         Deconvolve instrument response.
@@ -2780,6 +2780,10 @@ seismometer_correction_simulation.html#using-a-resp-file>`_.
         :type zero_mean: bool
         :param zero_mean: If `True`, the mean of the waveform data is
             subtracted in time domain prior to deconvolution.
+        :type fast: bool
+        :param fast: When set to True (default), then the calculation of the
+            response for a large number of frequencies is sped up through
+            interpolation. Relevant for traces with >10000 samples.
         :type taper: bool
         :param taper: If `True`, a cosine taper is applied to the waveform data
             in time domain prior to deconvolution.
