### 2020-07-14 61d5c12
add 'fast'-option to call
