### 2020-07-15 8464072
reword changelog a bit
