### 2020-07-15 d798789
reftek: make parts a bit easier to read
Co-authored-by: dsentinel <lloyd@passcal.nmt.edu>
