### 2020-07-15 6feba2b
reftek: simplify code as proposed by @dsentinel
