### 2020-07-15 0a1cd7e
reftek: refactoring to be nice to our future selves
add some comments, easier to read code
