### 2020-07-15 01fa2cb
reftek: refactoring, use slice instead of arange
