### 2020-07-15 7cb87a6
reftek: should fix reading on big endian systems, untested though..
