### 2020-07-16 2776601
Update obspy/io/reftek/core.py
Co-authored-by: dsentinel <lloyd@passcal.nmt.edu>
