+++ .github/ISSUE_TEMPLATE.md
@@ -10,7 +10,7 @@ Before submitting an Issue, please review the [Issue Guidelines](https://github.
 * Please provide the following information:
   -  ObsPy version, Python version and Platform (Windows, OSX, Linux ...)
   -  How did you install ObsPy and Python (pip, anaconda, from source ...)
-  -  If possible please supply a short, self contained, correct example that
+  -  If possible please supply a [short, self contained, correct example](http://sscce.org/) that
      demonstrates the issue.
   -  If this is a regression (used to work in an earlier version of ObsPy),
      please note when it used to work.
