### 2020-07-30 e064248
pha: add plugin entry
