+++ CHANGELOG.txt
@@ -7,6 +7,8 @@ Changes:
      instrument response (see #2677)
  - obspy.clients.filesystem:
    * add get_waveforms_bulk() method to SDS client (see #2616, #2626)
+ - obspy.io.hypodd
+   * add PHA write support
  - obspy.io.reftek:
    * enable reading data with floating point sampling rates like low sampling
      rate state-of-health channels (see #2678)
