### 2020-07-30 3b31719
add read support for horizontal and vertical origin uncertainty
