### 2020-07-30 99e7322
pha: add warnings, allow writing of errors and weights
