### 2020-07-30 obspy/obspy#2688 @megies
gh actions: add workflows on master (on pushes)
Following up on these comments in #2591: https://github.com/obspy/obspy/pull/2591#issuecomment-666370888

---

`validate` workflow is targeting PRs exclusively. @d-chambers could you please make two more workflows that run on pushes to master?
 - one with full build matrix, only doing non-network tests
 - one with only one build, but including network tests

Reason to have two is kinda some of the discussion in pyOpenSci/software-review#16
 - we want network tests run regularly in CI, to have them in our [test reports](tests.obspy.org)
 - people are confused by the red badge in README

That way we could use the more stable test result excluding volatile network tests in the README but still have network tests run in CI.
