### 2020-07-30 obspy/obspy#2688 @d-chambers
Sure, I should have time today or tomorrow to work on this
