### 2020-07-30 def714a
README: fix some links on some badges
