### 2020-07-30 c117f4a
README: RIP depsy, discontinued project long ago
