### 2020-07-30 2084c98
README: add liberapay badge
