### 2020-07-30 ee08b2d
Readme: add badge to mailing list
