### 2020-07-30 5fb723e
add pha write entry for doctest
