+++ .github/workflows/master_network_tests.yml
@@ -1,11 +1,18 @@
-name: calculate coverage
+# This workflow runs anytime new commits are added to master branch.
+# It runs all network tests on the master branch using linux and one
+# python version.
+name: master_network_tests
 on:
   push:
     branches:
       - master
 
+# Set globally scoped environmental variables.
+env:
+  OBSPY_GITHUB_BRANCH: gh-actions
+
 jobs:
-  calc_coverage:
+  network_tests:
     runs-on: ubuntu-latest
 
     steps:
@@ -18,17 +25,23 @@ jobs:
           python-version: 3.7
           activate-environment: test
           environment-file: .github/test_conda_env.yml
-          condarc-file: .github/test_condarc.yml
 
       - name: install obspy
         shell: bash -l {0}
         run: |
           pip install -e .
 
+      - name: install obspy github api
+        shell: bash -l {0}
+        run: |
+          pip install git+https://github.com/obspy/obspy_github_api@${OBSPY_GITHUB_BRANCH}
+
       - name: run test suite
         shell: bash -l {0}
         run: |
-          coverage run --rcfile=.coveragerc -m obspy.scripts.runtests --no-flake8 -n gh-actions -r
+          export MODULELIST=`obshub get-module-list --group all --sep ','`
+          export MODULELISTSPACES=`obshub get-module-list --group all --sep ' '`
+          coverage run --rcfile=.coveragerc --source=${MODULELIST} -m obspy.scripts.runtests --no-flake8 -n gh-actions -r --ci-url="${CI_URL}" $MODULELISTSPACES
           coverage xml -o coverage.xml
 
       - name: upload coverage
