+++ .github/workflows/pull_request.yml
@@ -12,7 +12,7 @@ env:
   PR_URL: https://github.com/obspy/obspy/pull/${{ github.event.pull_request.number }}
   CI_URL: https://github.com/obspy/obspy/actions/runs/${{ github.run_id }}
   CONFIG_PATH: obspy_ci_config/conf.json
-  GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}F
+  GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
   OBSPY_GITHUB_BRANCH: gh-actions
 
 
