### 2020-07-31 cf76420
update obspy_github_api release version
