+++ .github/workflows/master_network_tests.yml
@@ -9,7 +9,7 @@ on:
 
 # Set globally scoped environmental variables.
 env:
-  OBSPY_GITHUB_BRANCH: gh-actions
+  OBSPY_GITHUB_BRANCH: release_0.10.0
 
 jobs:
   network_tests:
