### 2020-07-31 06ff2cc
fix module collection for obspy test runner
