### 2020-08-03 3caf62c
Fix #2675, consistent deep and shallow copies for inventories
