### 2020-08-03 e0beb5c
Fix documentation of Inventory.extend
