### 2020-08-03 e99bc46
Combine Inventory metadata in Inventory.extend
