### 2020-08-03 40715fb
pha: add test using eventid_map in write call
