### 2020-08-03 d90467f
pha: add warning for very large catalog
