### 2020-08-03 1c327ff
pha: add info about which events were being skipped
Co-authored-by: Calum Chamberlain <calum.chamberlain@vuw.ac.nz>
