### 2020-08-03 d422c47
pha: add minimal write test, allow None in several fields
