### 2020-08-04 35f19e7
move _seed_id_map helper function to inventory.util
