+++ obspy/io/sh/evt.py
@@ -19,7 +19,7 @@ from obspy.core.event import (Arrival, Catalog, Event,
                               OriginUncertainty, Pick, ResourceIdentifier,
                               StationMagnitude, WaveformStreamID)
 from obspy.core.event.header import EvaluationMode, EventType, PickOnset
-from obspy.core.util.misc import _seed_id_map
+from obspy.core.inventory.util import _seed_id_map
 from obspy.io.sh.core import to_utcdatetime
 
 
