### 2020-08-04 fdce77c
fix nlloc test case (channel was not written to waveformid)
