### 2020-08-04 8681b9e
refactor nlloc resolve_seedid
