### 2020-08-04 7c9ff5e
allow unused kwargs, add more tests
