### 2020-08-04 7934ca2
add more tests, preserve ? as component
