### 2020-08-04 5acea4d
add support to resolve seed ids for nordic
