### 2020-08-04 4979195
refactor sh resolve_seedid
