+++ obspy/io/sh/evt.py
@@ -19,7 +19,7 @@ from obspy.core.event import (Arrival, Catalog, Event,
                               OriginUncertainty, Pick, ResourceIdentifier,
                               StationMagnitude, WaveformStreamID)
 from obspy.core.event.header import EvaluationMode, EventType, PickOnset
-from obspy.core.inventory.util import _seed_id_map
+from obspy.core.inventory.util import _resolve_seedid, _add_resolve_seedid_doc
 from obspy.io.sh.core import to_utcdatetime
 
 
@@ -136,8 +136,8 @@ def _mags(obj, evid, stamag=False, wid=None):
     return mags, pm
 
 
-def _read_evt(filename, inventory=None, id_map=None, id_default='.{}..{}',
-              encoding='utf-8'):
+@_add_resolve_seedid_doc
+def _read_evt(filename, encoding='utf-8', **kwargs):
     """
     Read a SeismicHandler EVT file and returns an ObsPy Catalog object.
 
@@ -146,28 +146,6 @@ def _read_evt(filename, inventory=None, id_map=None, id_default='.{}..{}',
         ObsPy :func:`~obspy.core.event.read_events` function, call this
         instead.
 
-    The optional parameters all deal with the problem, that the EVT format
-    only stores station names for the picks, but the Pick object expects
-    a SEED id. A SEED id template is retrieved for each station by the
-    following procedure:
-
-    1. look at id_map for a direct station name match and use the specified
-       template
-    2. if 1 did not succeed, look if the station is present in inventory and
-       use its first channel as template
-    3. if 1 and 2 did not succeed, use specified default template (id_default)
-
-    :param str filename: File or file-like object in text mode.
-    :type inventory: :class:`~obspy.core.inventory.inventory.Inventory`
-    :param inventory: Inventory used to retrieve network code, location code
-        and channel code of stations (SEED id).
-    :param dict id_map: Default templates for each station
-        (example: `id_map={'MOX': 'GR.{}..HH{}'`).
-        The values must contain three dots and two `{}` which are
-        substituted by station code and component.
-    :param str id_default: Default SEED id template.
-        The value must contain three dots and two `{}` which are
-        substituted by station code and component.
     :param str encoding: encoding used (default: utf-8)
 
     :rtype: :class:`~obspy.core.event.Catalog`
@@ -178,7 +156,6 @@ def _read_evt(filename, inventory=None, id_map=None, id_default='.{}..{}',
 
         Compare with http://www.seismic-handler.org/wiki/ShmDocFileEvt
     """
-    seed_map = _seed_id_map(inventory, id_map)
     with io.open(filename, 'r', encoding=encoding) as f:
         temp = f.read()
     # first create phases and phases_o dictionaries for different phases
@@ -223,9 +200,11 @@ def _read_evt(filename, inventory=None, id_map=None, id_default='.{}..{}',
         for p in phases[evid]:
             sta = p.get('station code', '')
             comp = p.get('component', '')
-            wid = seed_map.get(sta, id_default)
-            wid = WaveformStreamID(seed_string=wid.format(sta, comp))
-            pick = Pick(waveform_id=wid, **_kw(p, 'pick'))
+            pick_kwargs = _kw(p, 'pick')
+            widargs = _resolve_seedid(
+                    sta, comp, time=pick_kwargs['time'], **kwargs)
+            wid = WaveformStreamID(*widargs)
+            pick = Pick(waveform_id=wid, **pick_kwargs)
             arrival = Arrival(pick_id=pick.resource_id, **_kw(p, 'arrival'))
             picks.append(pick)
             arrivals.append(arrival)
