### 2020-08-04 cac323c
refactor pha resolve_seedid, more specific warnings
