### 2020-08-04 7c12ae6
remove obsolete _seed_id_map function
